% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{\%<+\%}
\alias{\%<+\%}
\title{\%<+\%}
\usage{
pg \%<+\% data
}
\arguments{
\item{pg}{ggplot2 object}

\item{data}{annotation data}
}
\value{
ggplot object with annotation data added
}
\description{
add annotation data to a tree
}
\author{
Yu Guangchuang
}

