% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-scale-color.R
\name{add_colorbar}
\alias{add_colorbar}
\title{add_colorbar}
\usage{
add_colorbar(p, color, x = NULL, ymin = NULL, ymax = NULL,
  font.size = 4)
}
\arguments{
\item{p}{tree view}

\item{color}{output of scale_color function}

\item{x}{x position}

\item{ymin}{ymin}

\item{ymax}{ymax}

\item{font.size}{font size}
}
\value{
ggplot2 object
}
\description{
add colorbar legend
}
\author{
Guangchuang Yu
}
