% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{add_legend}
\alias{add_legend}
\title{add_legend}
\usage{
add_legend(p, width = NULL, x = NULL, y = NULL, offset = NULL,
  font.size = 4, ...)
}
\arguments{
\item{p}{tree view}

\item{width}{width of legend}

\item{x}{x position}

\item{y}{y position}

\item{offset}{offset of text and line}

\item{font.size}{font size}

\item{...}{additional parameter}
}
\value{
tree view
}
\description{
add evolution distance legend
}
\author{
Guangchuang Yu
}

