% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{aes}
\alias{aes}
\title{creates a lists of unevaluated expressions}
\source{
This is just the imported function
from the ggplot2 package. The documentation you should
read for the aes function can be found here: \link[ggplot2]{aes}
}
\usage{
aes(x, y, ...)
}
\arguments{
\item{x}{name values}

\item{y}{name values}

\item{...}{additional name values}
}
\description{
creates a lists of unevaluated expressions
}
\seealso{
\link[ggplot2]{aes}
}

