% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x-axis.R
\name{align_x}
\alias{align_x}
\alias{align_y}
\title{align_x}
\usage{
align_x(gg, expand_limits = expand_scale(0, 0.6), by = "x")

align_y(gg, expand_limits = expand_scale(0, 0.6), by = "y")
}
\arguments{
\item{gg}{ggplot object}

\item{expand_limits}{amount to expand the limits}

\item{by}{one of 'x' or 'y'}
}
\value{
ggplot2 object with new limits and expand
}
\description{
set axis limits (x or y specific by 'by' parameter) of a ggplot object (left hand side)
based on the x (align_x function) or y (align_y function) limits of another ggplot object (right hand side).
This is useful for using cowplot or patchwork to align ggplot objects.
}
\author{
Guangchuang Yu
}
