% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{annotation_clade}
\alias{annotation_clade}
\title{annotation_clade}
\usage{
annotation_clade(tree_view, node, label, bar.size = 2, font.size = 4,
  offset = 0, offset.text = NULL, ...)
}
\arguments{
\item{tree_view}{tree view}

\item{node}{node number}

\item{label}{clade label}

\item{bar.size}{bar size}

\item{font.size}{font size}

\item{offset}{offset of bar from the tree}

\item{offset.text}{offset of label from bar}

\item{...}{additional parameter}
}
\value{
ggplot2
}
\description{
annotate a selected clade with internal node number
}
\author{
Guangchuang Yu
}

