% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{annotation_clade2}
\alias{annotation_clade2}
\title{annotation_clade2}
\usage{
annotation_clade2(tree_view, tip1, tip2, label, bar.size = 2, font.size = 4,
  offset = 0, offset.text = NULL, ...)
}
\arguments{
\item{tree_view}{tree view}

\item{tip1}{tip1 label or id}

\item{tip2}{tip2 label or id}

\item{label}{clade label}

\item{bar.size}{bar size}

\item{font.size}{font size}

\item{offset}{offset of bar from the tree}

\item{offset.text}{offset of label from bar}

\item{...}{additional parameter}
}
\value{
ggplot2
}
\description{
annotate a clade with selected upper and lower tips
}
\author{
Guangchuang Yu
}

