% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylopic.R
\name{annotation_image}
\alias{annotation_image}
\title{annotation_image}
\usage{
annotation_image(tree_view, img_info, width = 0.1, align = TRUE,
  linetype = "dotted", linesize = 1, offset = 0)
}
\arguments{
\item{tree_view}{tree view}

\item{img_info}{data.frame with first column of taxa name and second column of image names}

\item{width}{width of the image to be plotted in image}

\item{align}{logical}

\item{linetype}{line type if align = TRUE}

\item{linesize}{line size if align = TRUE}

\item{offset}{offset of image from the tree view}
}
\value{
tree view
}
\description{
annotation taxa with images
}
\author{
Guangchuang Yu
}

