% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{apeBootstrap-class}
\alias{apeBootstrap-class}
\title{Class "apeBootstrap"
This class stores ape bootstrapping analysis result}
\description{
Class "apeBootstrap"
This class stores ape bootstrapping analysis result
}
\section{Slots}{

\describe{
\item{\code{phylo}}{phylo object of treetext}

\item{\code{fields}}{available features}

\item{\code{bootstrap}}{bootstrap value}

\item{\code{extraInfo}}{extra information}
}}
\author{
Guangchuang Yu \url{http://guangchuangyu.github.io}
}
\keyword{classes}

