% Generated by roxygen2 (4.1.0): do not edit by hand
\name{as.binary.phylo}
\alias{as.binary.phylo}
\title{as.binary}
\usage{
\method{as.binary}{phylo}(tree, ...)
}
\arguments{
\item{tree}{phylo, object}

\item{...}{additional parameter}
}
\value{
binary tree
}
\description{
convert polytomy to binary tree
}
\details{
as.binary method for \code{phylo} object
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

