% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/treeIO.R
\name{as.data.frame.phylo}
\alias{as.data.frame.phylo}
\title{as.data.frame}
\usage{
\method{as.data.frame}{phylo}(x, ...)
}
\arguments{
\item{x}{phylo object}

\item{...}{additional parameter}
}
\value{
data.frame
}
\description{
as.data.frame
}
\author{
Yu Guangchuang
}

