% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Date.R
\name{as.Date.decimal}
\alias{as.Date.decimal}
\title{as.Date.decimal}
\usage{
\method{as.Date}{decimal}(x)
}
\arguments{
\item{x}{numerical number, eg 2014.34}
}
\value{
Date
}
\description{
convert decimal format to Date, eg "2014.34" to "2014-05-05"
}
\author{
Guangchuang Yu
}

