% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.polytomy.R
\name{as.polytomy}
\alias{as.polytomy}
\title{as.polytomy}
\usage{
as.polytomy(tree, feature, fun)
}
\arguments{
\item{tree}{tree object}

\item{feature}{selected feature}

\item{fun}{function to select nodes to collapse}
}
\value{
polytomy tree
}
\description{
collapse binary tree to polytomy by applying 'fun' to 'feature'
}
\author{
Guangchuang
}
