% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/beast.R
\docType{class}
\name{beast-class}
\alias{beast-class}
\alias{set.treeinfo,beast-method}
\alias{set.treeinfo<-,beast-method}
\title{Class "beast"
This class stores information of beast output}
\description{
Class "beast"
This class stores information of beast output
}
\section{Slots}{

\describe{
\item{\code{fields}}{beast statistic variables}

\item{\code{treetext}}{tree text in beast file}

\item{\code{tree}}{tree phylo object}

\item{\code{translation}}{tip number to name translation in beast file}

\item{\code{stats}}{beast statistics}

\item{\code{treeinfo}}{tree information for plotting}

\item{\code{file}}{beast file, nexus format}
}}
\author{
Guangchuang Yu \url{http://ygc.name}
}
\seealso{
\code{\link{show}} \code{\link{get.fields}}
          \code{\link{ggtree}}
}
\keyword{classes}

