% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{codeml-class}
\alias{codeml-class}
\title{Class "codeml"
This class stores information of output from codeml}
\description{
Class "codeml"
This class stores information of output from codeml
}
\section{Slots}{

\describe{
\item{\code{mlc}}{A \code{code_mlc} object}

\item{\code{rst}}{A \code{paml_rst} object}

\item{\code{extraInfo}}{extra information}
}}
\seealso{
\linkS4class{codeml_mlc} \linkS4class{paml_rst}
}
\keyword{codeml}

