% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{codeml_mlc-class}
\alias{codeml_mlc-class}
\title{Class "codeml_mlc"
This class stores information of mlc file frm codeml output}
\description{
Class "codeml_mlc"
This class stores information of mlc file frm codeml output
}
\section{Slots}{

\describe{
\item{\code{fields}}{available features}

\item{\code{treetext}}{tree text}

\item{\code{phylo}}{phylo object}

\item{\code{dNdS}}{dN dS information}

\item{\code{mlcfile}}{mlc file}

\item{\code{extraInfo}}{extra information}
}}
\author{
Guangchuang Yu
}
\seealso{
\linkS4class{paml_rst} \linkS4class{codeml}
}
\keyword{classes}

