% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{collapse_clade}
\alias{collapse_clade}
\title{collapse_clade}
\usage{
collapse_clade(tree_view, node)
}
\arguments{
\item{tree_view}{tree view}

\item{node}{clade node}
}
\value{
tree view
}
\description{
collapse a clade
}
\author{
Guangchuang Yu
}

