% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date.R
\name{Date2decimal}
\alias{Date2decimal}
\title{Date2decimal}
\usage{
Date2decimal(x)
}
\arguments{
\item{x}{Date}
}
\value{
numeric
}
\description{
convert Date to decimal format, eg "2014-05-05" to "2014.34"
}
\author{
Guangchuang Yu
}

