% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date.R
\name{decimal2Date}
\alias{decimal2Date}
\title{decimal2Date}
\usage{
decimal2Date(x)
}
\arguments{
\item{x}{numerical number, eg 2014.34}
}
\value{
Date
}
\description{
convert decimal format to Date, eg "2014.34" to "2014-05-05"
}
\author{
Guangchuang Yu
}

