% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylopic.R
\name{download.phylopic}
\alias{download.phylopic}
\title{download.phylopic}
\usage{
download.phylopic(id, size = 512, color = "black", alpha = 1)
}
\arguments{
\item{id}{phyopic id}

\item{size}{size of phylopic}

\item{color}{color}

\item{alpha}{alpha}
}
\value{
matrix
}
\description{
download phylopic
}
\author{
Guangchuang Yu
}
