% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{expand_clade}
\alias{expand_clade}
\title{expand_clade}
\usage{
expand_clade(tree_view, node)
}
\arguments{
\item{tree_view}{tree view}

\item{node}{clade node}
}
\value{
tree view
}
\description{
expand collased clade
}
\author{
Guangchuang Yu
}

