% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlim.R
\name{expand_xlim}
\alias{expand_xlim}
\title{expand_xlim}
\usage{
expand_xlim(xlim, panel)
}
\arguments{
\item{xlim}{xlim}

\item{panel}{panel}
}
\value{
updated tree view
}
\description{
expand x axis limits for specific panel
}
\author{
guangchuang yu
}

