% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_plot.R
\name{facet_data}
\alias{facet_data}
\title{facet_data}
\usage{
facet_data(tree_view, panel)
}
\arguments{
\item{tree_view}{ggtree object}

\item{panel}{data plotted in specific panel. If only one dataset used in the panel, return the data frame, else return a list of data frames.}
}
\value{
data frame or a list of data frames
}
\description{
extract data used in \code{facet_plot} or \code{geom_facet}
}
\references{
G Yu, TTY Lam, H Zhu, Y Guan (2018). Two methods for mapping and visualizing associated data
on phylogeny using ggtree. Molecular Biology and Evolution, 35(2):3041-3043.
\url{https://doi.org/10.1093/molbev/msy194}
}
\author{
Guangchuang Yu
}
