% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-utilities.R
\name{facet_labeller}
\alias{facet_labeller}
\title{facet_labeller}
\usage{
facet_labeller(p, label)
}
\arguments{
\item{p}{facet_plot output}

\item{label}{labels of facet panels}
}
\value{
ggplot object
}
\description{
label facet_plot output
}
\author{
Guangchuang Yu
}
