% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_plot.R
\name{facet_plot}
\alias{facet_plot}
\title{facet_plot}
\usage{
facet_plot(p, panel, data, geom, mapping = NULL, ...)
}
\arguments{
\item{p}{tree view}

\item{panel}{panel name for plot of input data}

\item{data}{data to plot by 'geom', first column should be matched with tip label of tree}

\item{geom}{geom function to plot the data}

\item{mapping}{aes mapping for 'geom'}

\item{...}{additional parameters for 'geom'}
}
\value{
ggplot object
}
\description{
plot tree associated data in an additional panel
}
\author{
Guangchuang Yu
}
