% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_plot.R
\name{facet_widths}
\alias{facet_widths}
\title{facet_widths}
\usage{
facet_widths(p, widths)
}
\arguments{
\item{p}{ggplot or ggtree object}

\item{widths}{relative widths of facet panels}
}
\value{
ggplot object by redrawing the figure (not a modified version of input object)
}
\description{
set relative widths (for column only) of facet plots
}
\author{
Guangchuang Yu
}
