% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.phylo}
\alias{fortify.phylo}
\title{fortify}
\usage{
\method{fortify}{phylo}(model, data, ...)
}
\arguments{
\item{model}{phylo object}

\item{data}{not use here}

\item{...}{additional parameter}
}
\value{
data.frame
}
\description{
fortify
}
\author{
Yu Guangchuang
}

