% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_segment.R
\name{geom_aline}
\alias{geom_aline}
\title{geom_aline}
\usage{
geom_aline(mapping = NULL, linetype = "dotted", linewidth = 1, ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{linetype}{set line type of the line, defaults to "dotted"}

\item{linewidth}{set width of the line, defaults to 1}

\item{...}{additional parameter}
}
\value{
aline layer
}
\description{
add horizontal align lines layer to a tree
}
\details{
'geom_aline'align all tips to the longest one by adding
padding characters to the right side of the tip.
}
\author{
Yu Guangchuang
}
