% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cladelabel2.R
\name{geom_cladelabel2}
\alias{geom_cladelabel2}
\title{geom_cladelabel2}
\usage{
geom_cladelabel2(
  node,
  label,
  offset = 0,
  offset.text = 0,
  offset.bar = 0,
  align = FALSE,
  barsize = 0.5,
  fontsize = 3.88,
  hjust = 0,
  geom = "text",
  color = NULL,
  family = "sans",
  parse = FALSE,
  horizontal = TRUE,
  ...
)
}
\arguments{
\item{node}{selected node}

\item{label}{clade label}

\item{offset}{offset of bar and text from the clade}

\item{offset.text}{offset of text from bar}

\item{offset.bar}{offset of bar from text}

\item{align}{logical}

\item{barsize}{size of bar}

\item{fontsize}{size of text}

\item{hjust}{justify text horizontally}

\item{geom}{one of 'text' or 'label'}

\item{color}{color for clade & label, of length 1 or 2}

\item{family}{sans by default, can be any supported font}

\item{parse}{logical, whether parse label}

\item{horizontal}{logical, whether set label to horizontal,
defaults to TRUE.}

\item{...}{additional parameter}
}
\value{
ggplot layers
}
\description{
annotate a clade with bar and text label
}
\seealso{
\link{geom_cladelabel}
}
\author{
JustGitting
}
