% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hilight_encircle.R
\name{geom_hilight_encircle}
\alias{geom_hilight_encircle}
\title{geom_hilight_encircle}
\usage{
geom_hilight_encircle(data = NULL, node, mapping = NULL,
  fill = "steelblue", alpha = 0.5, expand = 0, ...)
}
\arguments{
\item{node}{selected node to hilight (required)}

\item{fill}{colour fill (default = steelblue)}

\item{alpha}{alpha (transparency) (default = 0.5)}

\item{expand}{expands the xspline clade region only (default = 0)}

\item{spread}{spread of shape? (default = 0.1)}

\item{linetype}{Line type of xspline (default = 1)}

\item{size}{Size of xspline line (default = 1)}

\item{s_shape}{Corresponds to shape of xspline (default = 0.5)}

\item{s_open}{Boolean switch determines if xspline shape is open or closed. (default = FALSE)}
}
\value{
ggplot2
}
\description{
layer of hilight clade with xspline
}
