% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_function.R
\name{geom_motif}
\alias{geom_motif}
\title{geom_motif}
\usage{
geom_motif(mapping, data, on, label, align = "centre", ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{on}{gene to center (i.e. set middle position of the \code{on} gene to 0)}

\item{label}{specify a column to be used to label genes}

\item{align}{where to place gene label, default is 'centre' and can be set to 'left' and 'right'}

\item{...}{additional parameters}
}
\value{
geom layer
}
\description{
geom layer to draw aligned motif
}
\author{
Guangchuang Yu
}
