% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{geom_nodelab2}
\alias{geom_nodelab2}
\title{geom_nodelab2}
\usage{
geom_nodelab2(
  mapping = NULL,
  nudge_x = 0,
  nudge_y = 0,
  geom = "text",
  hjust = 0.5,
  ...
)
}
\arguments{
\item{mapping}{aesthetic mappings, defaults to NULL}

\item{nudge_x}{horizontal adjustment to nudge labels, defaults to 0}

\item{nudge_y}{vertical adjustment to nudge labels, defaults to 0}

\item{geom}{one of 'text', "shadowtext", 'label', 'image' and 'phylopic'}

\item{hjust}{horizontal alignment, defaults to 0.5}

\item{...}{additional parameters, see also
the additional parameters of \code{\link[=geom_tiplab]{geom_tiplab()}}.}
}
\value{
node label layer
}
\description{
add node label for circular layout
}
\author{
Guangchuang Yu
}
