% Generated by roxygen2 (4.1.0): do not edit by hand
\name{geom_nplace}
\alias{geom_nplace}
\title{geom_nplace}
\usage{
geom_nplace(data, map, place, ...)
}
\arguments{
\item{data}{placement data.frame}

\item{map}{edge column name}

\item{place}{place info}

\item{...}{additional parameter}
}
\value{
text layer
}
\description{
add placement based on node
}
\author{
ygc
}

