% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{geom_phylopic}
\alias{geom_phylopic}
\title{geom_phylopic}
\usage{
geom_phylopic(phylopic_id, x = NULL, y = NULL, width = NULL, size = 512,
  color = "black", alpha = 1)
}
\arguments{
\item{phylopic_id}{phylopic id}

\item{x}{x position}

\item{y}{y position}

\item{width}{width of phylopic}

\item{size}{size of phylopic to download}

\item{color}{color}

\item{alpha}{alpha}
}
\value{
phylopic layer
}
\description{
add phylopic layer
}
\author{
Guangchuang Yu
}

