% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{geom_place}
\alias{geom_place}
\title{geom_place}
\usage{
geom_place(data, map, place, by = "node", ...)
}
\arguments{
\item{data}{data}

\item{map}{mapping variable}

\item{place}{placement info}

\item{by}{one of "node" and "edge"}

\item{...}{additional parameter}
}
\value{
text layer
}
\description{
place annotation in tree
}
\examples{
nwk <- system.file("extdata", "sample.nwk", package="ggtree")
library(ape)
tree <- read.tree(nwk)
dd <- data.frame(taxa=sample(LETTERS[1:13], 6), value=round(abs(rnorm(6))*100))
ggtree(tree) + geom_tiplab() +
  geom_place(data=dd, map="taxa", place="value",
             hjust=0.8, vjust=-0.4, size=4, color="red")
}
\author{
ygc
}

