% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_segment.R
\name{geom_segment2}
\alias{geom_segment2}
\title{geom_segment2}
\usage{
geom_segment2(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  lineend = "butt",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  nudge_x = 0,
  arrow = NULL,
  arrow.fill = NULL,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings, defaults to NULL}

\item{data}{A layer specific dataset -
only needed if you want to override the plot defaults.}

\item{stat}{Name of stat to modify data.}

\item{position}{The position adjustment to use for overlapping points on this layer.}

\item{lineend}{Line end style, one of butt (default), round and square.}

\item{na.rm}{If "FALSE" (default), missing values are removed with a warning. If "TRUE", missing values are silently removed, logical.}

\item{show.legend}{Whether to show legend, logical.}

\item{inherit.aes}{Whether to inherit aesthetic mappings, logical, defaults to "TRUE".}

\item{nudge_x}{adjust the horizontal position of the segments.}

\item{arrow}{specification for arrow heads, as created by arrow().}

\item{arrow.fill}{fill color to usse for the arrow head (if closed). \code{NULL} means use \code{colour} aesthetic.}

\item{...}{additional parameter}
}
\value{
add segment layer
}
\description{
geom_segment2 support aes(subset) via setup_data
}
\details{
'geom_segment2' is a modified version of geom_segment, with subset aesthetic supported
}
\seealso{
\link[ggplot2:geom_segment]{geom_segment}
}
\author{
Guangchuang Yu
}
