% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{geom_text}
\alias{geom_text}
\title{text annotations}
\source{
This is just the imported function
from the ggplot2 package. The documentation you should
read for the geom_text function can be found here: \link[ggplot2]{geom_text}
}
\usage{
geom_text(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", parse = FALSE, show.legend = NA,
  inherit.aes = TRUE, ..., nudge_x = 0, nudge_y = 0,
  check_overlap = FALSE)
}
\arguments{
\item{mapping}{the aesthetic mapping}

\item{data}{A layer specific dataset -
only needed if you want to override he plot defaults.}

\item{stat}{The statistical transformation to use on the data for this layer}

\item{position}{The position adjustment to use for overlapping points on this layer}

\item{parse}{if TRUE, the labels will be passd into expressions}

\item{show.legend}{logical}

\item{inherit.aes}{logical}

\item{...}{other arguments passed on to 'layer'}

\item{nudge_x}{horizontal adjustment}

\item{nudge_y}{vertical adjustment}

\item{check_overlap}{if TRUE, text that overlaps previous text in the same layer will not be plotted}
}
\description{
text annotations
}
\seealso{
\link[ggplot2]{geom_text}
}

