% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_text.R
\name{geom_text2}
\alias{geom_text2}
\title{geom_text2}
\usage{
geom_text2(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", parse = FALSE, na.rm = TRUE, show.legend = NA,
  inherit.aes = TRUE, ..., nudge_x = 0, nudge_y = 0,
  check_overlap = FALSE)
}
\value{
text layer
}
\description{
geom_text2 support aes(subset) via setup_data
}
\author{
Guangchuang Yu
}
\seealso{
\link[ggplot2]{geom_text}
}

