% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{geom_tippoint}
\alias{geom_tippoint}
\title{geom_tippoint}
\usage{
geom_tippoint(...)
}
\arguments{
\item{...}{additional parameter}
}
\value{
tip point layer
}
\description{
add points layer of tips
}
\examples{
require(ape)
tr <- rtree(10)
ggtree(tr) + geom_tippoint()
}
\author{
Yu Guangchuang
}

