% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tree.R
\name{geom_tree2}
\alias{geom_tree2}
\title{geom_tree2}
\usage{
geom_tree2(layout = "rectangular", ...)
}
\arguments{
\item{layout}{one of 'rectangular', 'slanted', 'circular', 'radial' or 'unrooted'}

\item{...}{additional parameter}
}
\value{
tree layer
}
\description{
add tree layer
}
\author{
Yu Guangchuang
}

