% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_treescale.R
\name{geom_treescale}
\alias{geom_treescale}
\title{geom_treescale}
\usage{
geom_treescale(
  x = NULL,
  y = NULL,
  width = NULL,
  offset = NULL,
  offset.label = NULL,
  label = NULL,
  color = "black",
  linesize = 0.5,
  fontsize = 3.88,
  family = "sans"
)
}
\arguments{
\item{x}{set x position of the scale}

\item{y}{set y position of the scale}

\item{width}{set the length of the tree scale}

\item{offset}{set offset of text to line, defaults to NULL}

\item{offset.label}{set offset of the scale title to line.}

\item{label}{set the title of tree scale, defaults to NULL.}

\item{color}{set color of the scale}

\item{linesize}{set size of line}

\item{fontsize}{set size of text}

\item{family}{'sans' by default, can be any supported font}
}
\value{
ggplot layers
}
\description{
add tree scale to a tree
}
\details{
'geom_treescale' automatically adds a scale bar for evolutionary distance
}
\references{
For demonstration of this function, please refer to chapter 4.3.1 of
\emph{Data Integration, Manipulation and Visualization of Phylogenetic Trees}
\url{http://yulab-smu.top/treedata-book/index.html} by Guangchuang Yu.
}
\author{
Guangchuang Yu
}
