% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-zoom-clade.R
\name{geom_zoom_clade}
\alias{geom_zoom_clade}
\title{geom_zoom_clade}
\usage{
geom_zoom_clade(node)
}
\arguments{
\item{node}{internal node number}
}
\value{
updated tree view
}
\description{
zoom selected clade of a tree
}
\author{
Guangchuang Yu
}
