% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylopic.R
\name{get.phylopic}
\alias{get.phylopic}
\title{get.phylopic}
\usage{
get.phylopic(id, size = 512, color = "black", alpha = 1)
}
\arguments{
\item{id}{phylopic id}

\item{size}{size of the phylopic}

\item{color}{color}

\item{alpha}{alpha}
}
\value{
grob object
}
\description{
download phylopic and convert to grob object
}
\author{
Guangchuang Yu
}

