% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/jplace.R
\docType{methods}
\name{get.placements}
\alias{get.placements}
\alias{get.placements,jplace,ANY-method}
\title{get.placements method}
\usage{
get.placements(object, by, ...)

get.placements(object, by, ...)
}
\arguments{
\item{object}{jplace object}

\item{by}{get best hit or others}

\item{...}{additional parameter}
}
\value{
data.frame
}
\description{
get.placements method

get.placement method
}
\examples{
jp <- system.file("extdata", "sample.jplace", package="ggtree")
jp <- read.jplace(jp)
get.placements(jp, by="all")
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

