% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{get.treeinfo}
\alias{get.treeinfo}
\alias{get.treeinfo,jplace,ANY-method}
\title{get.treeinfo method}
\usage{
get.treeinfo(object, ladderize = TRUE, right = FALSE, ...)

get.treeinfo(object, ladderize, right, ...)
}
\arguments{
\item{object}{jplace object}

\item{ladderize}{ladderize, logical}

\item{right}{logical, parameter for ladderize}

\item{...}{additional parameter}
}
\value{
data.frame

data.frame
}
\description{
get.treeinfo method generics

get.treeinfo method
}
\author{
Guangchuang Yu \url{http://ygc.name}

Guangchuang Yu \url{http://ygc.name}
}

