% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/jplace.R
\docType{methods}
\name{get.treetext}
\alias{get.treetext}
\alias{get.treetext,jplace,ANY-method}
\title{get.treetext method}
\usage{
get.treetext(object, ...)

get.treetext(object, ...)
}
\arguments{
\item{object}{one of \code{jplace}, \code{beast} object}

\item{...}{additional parameter}
}
\value{
phylo object
}
\description{
get.treetext method

get.treetext method
}
\examples{
\dontrun{
jp <- system.file("extdata", "sample.jplace", package="ggtree")
jp <- read.jplace(jp)
get.treetext(jp)
}
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

