% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_balance.R
\name{get_balance_position}
\alias{get_balance_position}
\title{get_balance_position}
\usage{
get_balance_position(treeview, node, direction)
}
\arguments{
\item{treeview}{tree view}

\item{node}{selected node}

\item{direction}{either (1 for 'up' or 2 for 'down')}
}
\value{
data.frame
}
\description{
get position of balance (xmin, xmax, ymin, ymax)
}
\author{
Justin Silverman
}
