% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hilight.R
\name{get_clade_position}
\alias{get_clade_position}
\title{get_clade_position}
\usage{
get_clade_position(treeview, node)
}
\arguments{
\item{treeview}{tree view}

\item{node}{selected node}
}
\value{
data.frame
}
\description{
get position of clade (xmin, xmax, ymin, ymax)
}
\author{
Guangchuang Yu
}
