% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_function.R
\name{get_heatmap_column_position}
\alias{get_heatmap_column_position}
\title{get_heatmap_column_position}
\usage{
get_heatmap_column_position(treeview, by = "bottom")
}
\arguments{
\item{treeview}{output of `gheatmap`}

\item{by}{one of 'bottom' or 'top'}
}
\value{
data.frame
}
\description{
return a data.frame that contains position information
for labeling column names of heatmap produced by `gheatmap` function
}
\author{
Guangchuang Yu
}
