% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade-functions.R
\name{get_taxa_name}
\alias{get_taxa_name}
\title{get_taxa_name}
\usage{
get_taxa_name(tree_view = NULL, node = NULL)
}
\arguments{
\item{tree_view}{tree view (i.e. the ggtree object). If tree_view is NULL, the last ggplot object will be used.}

\item{node}{internal node number to specify a clade. If NULL, using the whole tree}
}
\value{
ordered taxa name vector
}
\description{
get taxa name of a selected node (or tree if node=NULL) sorted by their position in plotting
}
\details{
This function extract an ordered vector of the tips from selected clade or the whole tree
based on the ggtree() plot.
}
\examples{
tree <- rtree(30)
p <- ggtree(tree)
get_taxa_name(p)
}
\author{
Guangchuang Yu
}
