% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytree.R
\name{getChild.df}
\alias{getChild.df}
\title{getChild.df}
\usage{
getChild.df(df, node)
}
\arguments{
\item{df}{tree data.frame}

\item{node}{is the node id of child in tree.}
}
\value{
list of child node ids of parent
}
\description{
Get list of child node id numbers of parent node
}
