% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-utilities.R
\name{getNodeAngle.df}
\alias{getNodeAngle.df}
\title{getNodeAngle.df}
\usage{
getNodeAngle.df(df, origin_node_id, node_id)
}
\arguments{
\item{df}{tree data.frame}

\item{origin_node_id}{origin node id number}

\item{node_id}{end node id number}
}
\value{
angle in range \verb{[-1, 1]}, i.e. degrees/180, radians/pi
}
\description{
getNodeAngle.df
}
