% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytree.R
\name{getNodeEuclDistances}
\alias{getNodeEuclDistances}
\title{Get the distances from the node to all other nodes in data.frame (including itself if in df)}
\usage{
getNodeEuclDistances(df, node)
}
\description{
Get the distances from the node to all other nodes in data.frame (including itself if in df)
}
