% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-utilities.R
\name{getNodeNum}
\alias{getNodeNum}
\title{getNodeNum}
\usage{
getNodeNum(tr)
}
\arguments{
\item{tr}{phylo object}
}
\value{
number
}
\description{
calculate total number of nodes
}
\author{
Guangchuang Yu
}

