% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytree.R
\name{getNodesBreadthFirst.df}
\alias{getNodesBreadthFirst.df}
\title{getNodesBreadthFirst.df}
\usage{
getNodesBreadthFirst.df(df)
}
\arguments{
\item{df}{tree data.frame}
}
\value{
list of node id's in breadth-first order.
}
\description{
Get the nodes of tree from root in breadth-first order.
}
