% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytree.R
\name{getParent.df}
\alias{getParent.df}
\title{getParent.df}
\usage{
getParent.df(df, node)
}
\arguments{
\item{df}{tree data.frame}

\item{node}{is the node id of child in tree.}
}
\value{
integer node id of parent
}
\description{
Get parent node id of child node.
}
