% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytree.R
\name{GetSubtreeUnrooted}
\alias{GetSubtreeUnrooted}
\title{GetSubtreeUnrooted}
\usage{
GetSubtreeUnrooted(tree, node)
}
\arguments{
\item{tree}{ape phylo tree object}

\item{node}{is the tree node id from which the subtrees are derived.}
}
\value{
named list of subtrees with the root id of subtree and list of node id's making up subtree.
}
\description{
Get all subtrees of specified node. This includes all ancestors and relatives of node and
return named list of subtrees.
}
