% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytree.R
\name{getTreeArcAngles}
\alias{getTreeArcAngles}
\title{getTreeArcAngles}
\usage{
getTreeArcAngles(df, origin_id, subtree)
}
\arguments{
\item{df}{tree data.frame}

\item{origin_id}{node id from which to calculate left and right hand angles of subtree.}

\item{subtree}{named list of root id of subtree and list of node ids for given subtree.}
}
\value{
named list with right and left angles in range [0,2] i.e 1 = 180 degrees, 1.5 = 270 degrees.
}
\description{
Find the right (clockwise rotation, angle from +ve x-axis to furthest subtree nodes) and
left (anti-clockwise angle from +ve x-axis to subtree)
}

