% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{ggplotGrob}
\alias{ggplotGrob}
\title{generate a ggplot2 plot grob}
\source{
This is just the imported function
from the ggplot2 package. The documentation you should
read for the ggplotGrob function can be found here: \link[ggplot2]{ggplotGrob}
}
\usage{
ggplotGrob(x)
}
\arguments{
\item{x}{ggplot2 object}
}
\description{
generate a ggplot2 plot grob
}
\seealso{
\link[ggplot2]{ggplotGrob}
}

