% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{\%<\%}
\alias{\%<\%}
\title{\%<\%}
\usage{
pg \%<\% x
}
\arguments{
\item{pg}{ggtree object}

\item{x}{tree object}
}
\value{
updated ggplot object
}
\description{
update tree
}
\details{
This operator apply the visualization directives in ggtree object (lhs)
to visualize another tree object (rhs), that is similar to Format Painter.
}
\examples{
library("ggplot2")
nwk <- system.file("extdata", "sample.nwk", package="treeio")
tree <- read.tree(nwk)
p <- ggtree(tree) + geom_tippoint(color="#b5e521", alpha=1/4, size=10)
p \%<\% rtree(30)
}
\author{
Guangchuang Yu
}
