% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_balance.R, R/geom_hilight.R,
%   R/geom_hilight_encircle.R
\docType{data}
\name{StatBalance}
\alias{StatBalance}
\alias{StatHilight}
\alias{GeomHilight}
\title{StatBalance}
\description{
StatBalance

StatHilight

GeomHilight
}
\keyword{datasets}
