% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtree-package.R
\docType{package}
\name{ggtree-package}
\alias{ggtree-package}
\alias{package-ggtree}
\title{ggtree: an R package for visualization of tree and annotation data}
\description{
'ggtree' extends the 'ggplot2' plotting system which implemented the grammar of graphics. 'ggtree' is designed for visualization and annotation of phylogenetic trees and other tree-like structures with their annotation data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.amazon.com/Integration-Manipulation-Visualization-Phylogenetic-Computational-ebook/dp/B0B5NLZR1Z/} (book)
  \item \url{http://onlinelibrary.wiley.com/doi/10.1111/2041-210X.12628} (paper)
  \item Report bugs at \url{https://github.com/YuLab-SMU/ggtree/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]
  \item Tommy Tsan-Yuk Lam \email{tylam.tommy@gmail.com} [thesis advisor]
  \item Shuangbin Xu \email{xshuangbin@163.com} (\href{https://orcid.org/0000-0003-3513-5362}{ORCID})
}

Other contributors:
\itemize{
  \item Lin Li \email{williamlee2220@qq.com} [contributor]
  \item Bradley Jones \email{bjones@cfenet.ubc.ca} [contributor]
  \item Justin Silverman \email{jsilve24@gmail.com} [contributor]
  \item Watal M. Iwasaki \email{heavy.watal@gmail.com} [contributor]
  \item Yonghe Xia \email{xiayh17@gmail.com} [contributor]
  \item Ruizhu Huang \email{ruizhuRH@gmail.com} [contributor]
}

}
\keyword{internal}
