% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/beast.R, R/codeml.R, R/codeml_mlc.R, R/hyphy.R, R/jplace.R, R/paml_rst.R, R/phylo.R
\docType{methods}
\name{groupClade}
\alias{groupClade}
\alias{groupClade,beast-method}
\alias{groupClade,codeml-method}
\alias{groupClade,codeml_mlc-method}
\alias{groupClade,hyphy-method}
\alias{groupClade,jplace-method}
\alias{groupClade,paml_rst-method}
\alias{groupClade,phylo-method}
\title{groupClade method}
\usage{
groupClade(object, node)

\S4method{groupClade}{beast}(object, node)

\S4method{groupClade}{codeml}(object, node)

\S4method{groupClade}{codeml_mlc}(object, node)

\S4method{groupClade}{hyphy}(object, node)

\S4method{groupClade}{jplace}(object, node)

\S4method{groupClade}{paml_rst}(object, node)

\S4method{groupClade}{phylo}(object, node)
}
\arguments{
\item{object}{supported objects, including phylo, paml_rst,
codeml_mlc, codeml, jplace, beast, hyphy}

\item{node}{a internal node or a vector of internal nodes}
}
\value{
group index
}
\description{
groupClade method
}

