% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/beast.R, R/codeml.R, R/codeml_mlc.R, R/ggtree.R, R/hyphy.R, R/jplace.R, R/paml_rst.R, R/phylo.R
\docType{methods}
\name{groupOTU}
\alias{groupOTU}
\alias{groupOTU,beast-method}
\alias{groupOTU,codeml-method}
\alias{groupOTU,codeml_mlc-method}
\alias{groupOTU,gg-method}
\alias{groupOTU,ggplot-method}
\alias{groupOTU,hyphy-method}
\alias{groupOTU,jplace-method}
\alias{groupOTU,paml_rst-method}
\alias{groupOTU,phylo-method}
\title{groupOTU method}
\usage{
groupOTU(object, focus, group_name = "group")

\S4method{groupOTU}{beast}(object, focus, group_name = "group")

\S4method{groupOTU}{codeml}(object, focus, group_name = "group")

\S4method{groupOTU}{codeml_mlc}(object, focus, group_name = "group")

\S4method{groupOTU}{ggplot}(object, focus, group_name = "group")

\S4method{groupOTU}{gg}(object, focus, group_name = "group")

\S4method{groupOTU}{hyphy}(object, focus, group_name = "group")

\S4method{groupOTU}{jplace}(object, focus, group_name = "group")

\S4method{groupOTU}{paml_rst}(object, focus, group_name = "group")

\S4method{groupOTU}{phylo}(object, focus, group_name = "group")
}
\arguments{
\item{object}{supported objects, including phylo, paml_rst,
codeml_mlc, codeml, jplace, beast, hyphy}

\item{focus}{a vector of tip (label or number) or a list of tips.}

\item{group_name}{name of the group, 'group' by default}
}
\value{
group index
}
\description{
groupOTU method
}

