% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phylo.R
\name{groupOTU.phylo}
\alias{groupOTU.phylo}
\title{groupOTU.phylo}
\usage{
groupOTU.phylo(phy, focus)
}
\arguments{
\item{phy}{tree object}

\item{focus}{tip list}
}
\value{
cluster index
}
\description{
group OTU
}
\author{
ygc
}

