% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tree.R
\name{groupOTU}
\alias{groupOTU}
\title{groupOTU}
\usage{
groupOTU(phy, focus)
}
\arguments{
\item{phy}{tree object}

\item{focus}{tip list}
}
\value{
cluster index
}
\description{
group OTU
}
\author{
ygc
}

