% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/beast.R, R/codeml.R, R/codeml_mlc.R, R/hyphy.R, R/paml_rst.R, R/phylo.R
\docType{methods}
\name{gzoom}
\alias{gzoom}
\alias{gzoom,beast-method}
\alias{gzoom,codeml-method}
\alias{gzoom,codeml_mlc-method}
\alias{gzoom,hyphy-method}
\alias{gzoom,paml_rst-method}
\alias{gzoom,phylo-method}
\title{gzoom method}
\usage{
gzoom(object, focus, subtree = FALSE, widths = c(0.3, 0.7))

\S4method{gzoom}{beast}(object, focus, subtree = FALSE, widths = c(0.3,
  0.7))

\S4method{gzoom}{codeml}(object, focus, subtree = FALSE, widths = c(0.3,
  0.7))

\S4method{gzoom}{codeml_mlc}(object, focus, subtree = FALSE, widths = c(0.3,
  0.7))

\S4method{gzoom}{hyphy}(object, focus, subtree = FALSE, widths = c(0.3,
  0.7))

\S4method{gzoom}{paml_rst}(object, focus, subtree = FALSE, widths = c(0.3,
  0.7))

\S4method{gzoom}{phylo}(object, focus, subtree = FALSE, widths = c(0.3,
  0.7))
}
\arguments{
\item{object}{supported tree objects}

\item{focus}{selected tips}

\item{subtree}{logical}

\item{widths}{widths}
}
\value{
figure
}
\description{
gzoom method
}

