% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree.R
\name{gzoom.phylo}
\alias{gzoom.phylo}
\title{gzoom}
\usage{
gzoom.phylo(phy, focus, subtree = FALSE, widths = c(0.3, 0.7))
}
\arguments{
\item{phy}{phylo object}

\item{focus}{selected tips}

\item{subtree}{logical}

\item{widths}{widths}
}
\value{
a list of ggplot object
}
\description{
plots simultaneously a whole phylogenetic tree and a portion of it.
}
\examples{
require(ape)
data(chiroptera)
gzoom(chiroptera, grep("Plecotus", chiroptera$tip.label))
}
\author{
ygc
}

