% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{hexpand}
\alias{hexpand}
\title{hexpand}
\usage{
hexpand(ratio, direction = 1)
}
\arguments{
\item{ratio}{expand x limits by amount of xrange * ratio}

\item{direction}{expand x limit at right hand side if direction is 1, or left hand side if direction is -1}
}
\value{
hexpand object
}
\description{
expand xlim by ratio of x range
}
\author{
Guangchuang Yu
}
