% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{hyphy-class}
\alias{hyphy-class}
\title{Class "hyphy"
This class stores information of HYPHY output}
\description{
Class "hyphy"
This class stores information of HYPHY output
}
\section{Slots}{

\describe{
\item{\code{fields}}{available features}

\item{\code{treetext}}{tree text}

\item{\code{phylo}}{phylo object}

\item{\code{seq_type}}{one of "NT" and "AA"}

\item{\code{subs}}{sequence substitutions}

\item{\code{AA_subs}}{Amino acid sequence substitution}

\item{\code{ancseq}}{ancestral sequences}

\item{\code{tip_seq}}{tip sequences}

\item{\code{tip.fasfile}}{fasta file of tip sequences}

\item{\code{tree.file}}{tree file}

\item{\code{ancseq.file}}{ancestral sequence file, nexus format}

\item{\code{extraInfo}}{extra information}
}}
\author{
Guangchuang Yu \url{http://guangchuangyu.github.io}
}
\seealso{
\linkS4class{paml_rst}
}
\keyword{classes}

