% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-identify.R
\name{identify.gg}
\alias{identify.gg}
\title{identify}
\usage{
\method{identify}{gg}(x = last_plot(), col = "auto", ...)
}
\arguments{
\item{x}{tree view}

\item{col}{selected columns to extract. Default is "auto" which will select all columns for 'ggplot' object and 'node' column for 'ggtree' object}

\item{...}{additional parameters, normally ignored}
}
\value{
closest data point
}
\description{
identify node by interactive click
}
\author{
Guangchuang Yu
}
