% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset.R
\name{geom_inset}
\alias{geom_inset}
\alias{inset}
\title{geom_inset}
\usage{
geom_inset(
  insets,
  width = 0.1,
  height = 0.1,
  hjust = 0,
  vjust = 0,
  x = "node",
  reverse_x = FALSE,
  reverse_y = FALSE
)

inset(
  tree_view,
  insets,
  width,
  height,
  hjust = 0,
  vjust = 0,
  x = "node",
  reverse_x = FALSE,
  reverse_y = FALSE
)
}
\arguments{
\item{insets}{a list of ggplot objects, named by node number}

\item{width}{width of inset, relative to the range of x-axis}

\item{height}{height of inset, relative to the range of y-axis}

\item{hjust}{horizontal adjustment}

\item{vjust}{vertical adjustment}

\item{x}{x position, one of 'node' and 'branch'}

\item{reverse_x}{whether x axis was reversed by scale_x_reverse}

\item{reverse_y}{whether y axis was reversed by scale_y_reverse}

\item{tree_view}{tree view}
}
\value{
inset layer

tree view with insets
}
\description{
add subplots to tree

add insets in a tree
}
\author{
Guangchuang Yu

Guangchuang Yu
}
