% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytree.R
\name{is.ggtree}
\alias{is.ggtree}
\title{is.ggtree}
\usage{
is.ggtree(x)
}
\arguments{
\item{x}{object}
}
\value{
TRUE or FALSE
}
\description{
test whether input object is produced by ggtree function
}
\author{
guangchuang yu
}

