% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{jplace-class}
\alias{get.placements,jplace-method}
\alias{get.tree,jplace-method}
\alias{get.treeinfo,jplace-method}
\alias{jplace-class}
\alias{set.treeinfo,jplace-method}
\alias{set.treeinfo<-,jplace-method}
\alias{show,jplace-method}
\title{Class "jplace"
This class stores information of jplace file.}
\description{
Class "jplace"
This class stores information of jplace file.
}
\section{Slots}{

\describe{
\item{\code{fields}}{colnames of first variable of placements}

\item{\code{tree}}{tree text}

\item{\code{placements}}{placement information}

\item{\code{version}}{version}

\item{\code{metadata}}{metadata}

\item{\code{treeinfo}}{tree infomation}
}}
\author{
Guangchuang Yu \url{http://ygc.name}
}
\seealso{
\code{\link{show}} \code{\link{get.tree}}
}
\keyword{classes}

