% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytree.R
\name{layoutDaylight}
\alias{layoutDaylight}
\title{Equal daylight layout method for unrooted trees.}
\usage{
layoutDaylight(tree, branch.length)
}
\arguments{
\item{tree}{phylo object}

\item{branch.length}{set to 'none' for edge length of 1. Otherwise the phylogenetic tree edge length is used.}
}
\value{
tree as data.frame with equal angle layout.
}
\description{
#' @title
}
\references{
The following aglorithm aims to implement the vague description of the "Equal-daylight Algorithm"
in "Inferring Phylogenies" pp 582-584 by Joseph Felsenstein.

```
Leafs are subtrees with no children
Initialise tree using equal angle algorithm
tree_df = equal_angle(tree)

nodes = get list of nodes in tree_df breadth-first
nodes = remove tip nodes.

```
}

