% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytree.R
\name{layoutEqualAngle}
\alias{layoutEqualAngle}
\title{layoutEqualAngle}
\usage{
layoutEqualAngle(tree, branch.length)
}
\arguments{
\item{tree}{phylo object}

\item{branch.length}{set to 'none' for edge length of 1. Otherwise the phylogenetic tree edge length is used.}
}
\value{
tree as data.frame with equal angle layout.
}
\description{
'Equal-angle layout algorithm for unrooted trees'
}
\references{
"Inferring Phylogenies" by Joseph Felsenstein.
}
