% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/paml.R
\name{mask}
\alias{mask}
\title{mask}
\usage{
mask(tree_object, field, site, mask_site = FALSE)
}
\arguments{
\item{tree_object}{tree object}

\item{field}{selected field}

\item{site}{site}

\item{mask_site}{if TRUE, site will be masked.
if FALSE, selected site will not be masked, while other sites will be masked.}
}
\value{
updated tree object
}
\description{
site mask
}
\author{
Guangchuang Yu
}

