% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge_tree.R
\name{merge_tree}
\alias{merge_tree}
\title{merge_tree}
\usage{
merge_tree(obj1, obj2)
}
\arguments{
\item{obj1}{tree object 1}

\item{obj2}{tree object 2}
}
\value{
tree object
}
\description{
merge two tree object
}
\author{
Guangchuang Yu
}

