% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCA.R
\name{MRCA}
\alias{MRCA}
\title{MRCA}
\usage{
MRCA(obj, tip)
}
\arguments{
\item{obj}{supported tree object or ggplot object}

\item{tip}{a vector of mode numeric or character specifying the tips}
}
\value{
MRCA of two or more tips
}
\description{
Find Most Recent Common Ancestor among a vector of tips
}
\author{
Guangchuang Yu
}
