% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{nhx-class}
\alias{nhx-class}
\title{Class "nhx"
This class stores nhx tree}
\description{
Class "nhx"
This class stores nhx tree
}
\section{Slots}{

\describe{
\item{\code{file}}{input file}

\item{\code{fields}}{available feature}

\item{\code{phylo}}{phylo object}

\item{\code{nhx_tags}}{tag information in nhx file}

\item{\code{extraInfo}}{extra information}
}}
\author{
Guangchuang Yu \url{http://ygc.name}
}
\keyword{classes}

