% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset.R
\name{nodebar}
\alias{nodebar}
\title{nodebar}
\usage{
nodebar(data, cols, color, alpha = 1, position = "stack")
}
\arguments{
\item{data}{a data.frame of stats with an additional column of node number named "node"}

\item{cols}{columns of the data.frame that store the stats}

\item{color}{set color of bars}

\item{alpha}{set transparency of the charts}

\item{position}{position of bars, if 'stack' (default) make bars stacked atop one another, 'dodge' make them dodged side-to-side}
}
\value{
list of ggplot objects
}
\description{
generate a list of bar charts for results of ancestral state reconstruction
}
\author{
Guangchuang Yu
}
