% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytree.R
\name{nodeid}
\alias{nodeid}
\title{nodeid}
\usage{
nodeid(x, label)
}
\arguments{
\item{x}{tree object or graphic object return by ggtree}

\item{label}{tip or node label(s)}
}
\value{
internal node number
}
\description{
convert tip or node label(s) to internal node number
}
\author{
Guangchuang Yu
}
