% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset.R
\name{nodepie}
\alias{nodepie}
\title{nodepie}
\usage{
nodepie(data, cols, color, alpha = 1)
}
\arguments{
\item{data}{a data.frame of stats with an additional column of node number named "node"}

\item{cols}{columns of the data.frame that store the stats}

\item{color}{set color of bars}

\item{alpha}{set transparency of the charts}
}
\value{
list of ggplot objects
}
\description{
generate a list of pie charts for results of ancestral stat reconstruction
}
\author{
Guangchuang Yu
}
