% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{open_tree}
\alias{open_tree}
\title{open_tree}
\usage{
open_tree(treeview, angle)
}
\arguments{
\item{treeview}{tree view in rectangular/circular layout}

\item{angle}{open the tree at a specific angle}
}
\value{
updated tree view
}
\description{
transform a tree in either rectangular or circular layout into the fan layout
that opens with a specific angle
}
\examples{
tree <- rtree(15)
p <- ggtree(tree) + geom_tiplab()
open_tree(p, 180)
}
\author{
Guangchuang Yu
}
