% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{paml_rst-class}
\alias{paml_rst-class}
\alias{set.subs,paml_rst-method}
\alias{set.subs<-,paml_rst-method}
\title{Class "paml_rst"
This class stores information of rst file from PAML output}
\description{
Class "paml_rst"
This class stores information of rst file from PAML output
}
\section{Slots}{

\describe{
\item{\code{fields}}{availabel attributes}

\item{\code{treetext}}{tree text}

\item{\code{phylo}}{phylo object}

\item{\code{seq_type}}{one of "NT" and "AA"}

\item{\code{tip_seq}}{sequences of tips}

\item{\code{marginal_ancseq}}{Marginal reconstruction of ancestral sequences}

\item{\code{joint_ancseq}}{Joint reconstruction of ancestral sequences}

\item{\code{marginal_subs}}{sequence substitutions based on marginal_ancseq}

\item{\code{joint_subs}}{sequence substitutions based on joint_ancseq}

\item{\code{marginal_AA_subs}}{Amino acid sequence substitutions based on marginal_ancseq}

\item{\code{joint_AA_subs}}{Amino acid sequence substitutions based on joint_ancseq}

\item{\code{rstfile}}{rst file}

\item{\code{extraInfo}}{extra information}
}}
\author{
Guangchuang Yu \url{http://ygc.name}
}
\seealso{
\linkS4class{codeml} \linkS4class{codeml_mlc}
}
\keyword{classes}

