% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{phangorn-class}
\alias{phangorn-class}
\title{Class "phangorn"
This class stores ancestral sequences inferred from 'phangorn'}
\description{
Class "phangorn"
This class stores ancestral sequences inferred from 'phangorn'
}
\section{Slots}{

\describe{
\item{\code{fields}}{available attributes}

\item{\code{phylo}}{phylo object}

\item{\code{seq_type}}{one of "NT" and "AA"}

\item{\code{tip_seq}}{sequences of tips}

\item{\code{ancseq}}{ancenstral sequences}

\item{\code{subs}}{sequence substitution}

\item{\code{AA_subs}}{Amino acid sequence substitution}

\item{\code{extraInfo}}{extra information}
}}
\author{
Guangchuang Yu \url{http://guangchuangyu.github.io}
}
\seealso{
\linkS4class{paml_rst}
}
\keyword{classes}

