% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{phylip-class}
\alias{phylip-class}
\title{Class "phylip"
This class stores phylip tree(s)}
\description{
Class "phylip"
This class stores phylip tree(s)
}
\section{Slots}{

\describe{
\item{\code{file}}{input file}

\item{\code{fields}}{available feature}

\item{\code{phylo}}{phylo or multiPhylo}

\item{\code{ntree}}{number of trees}

\item{\code{sequence}}{sequences}

\item{\code{extraInfo}}{extra information}
}}
\author{
Guangchuang Yu
}
\keyword{classes}

