% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylopic.R
\name{phylopic}
\alias{phylopic}
\title{phylopic}
\usage{
phylopic(tree_view, phylopic_id, size = 512, color = "black", alpha = 0.5,
  node = NULL, x = NULL, y = NULL, width = 0.1)
}
\arguments{
\item{tree_view}{tree view}

\item{phylopic_id}{phylopic id}

\item{size}{size of phylopic to download}

\item{color}{color}

\item{alpha}{alpha}

\item{node}{selected node}

\item{x}{x position}

\item{y}{y position}

\item{width}{width of phylopic}
}
\value{
phylopic layer
}
\description{
add phylopic layer
}
\author{
Guangchuang Yu
}
