% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phangorn.R
\name{phyPML}
\alias{phyPML}
\title{treeAnno.pml}
\usage{
phyPML(pmlTree, type = "ml")
}
\arguments{
\item{pmlTree}{tree in pml object, output of phangorn::optim.pml}

\item{type}{one of 'ml' and 'bayes' for inferring ancestral sequences}
}
\value{
phangorn object
}
\description{
tree annotation of sequence substitution by comparing to parent node
}
\author{
Yu Guangchuang
}

