% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%>\%}
\alias{\%>\%}
\title{pipe}
\source{
This is just the imported function
from the marittr package. The documentation you should
read for %>% function can be found here: \link[magrittr]{%>%}
}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{left hand side}

\item{rhs}{right hand side}
}
\description{
pipe
}

