% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phangorn.R
\name{pmlToSeq}
\alias{pmlToSeq}
\title{pmlToSeq}
\usage{
pmlToSeq(pml, type = "ml", includeAncestor = TRUE)
}
\arguments{
\item{pml}{pml object}

\item{type}{one of "marginal", "ml", "bayes"}

\item{includeAncestor}{logical}
}
\value{
XStringSet
}
\description{
convert pml object to XStringSet object
}
\author{
ygc
}

