% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-print.R
\name{print.beastList}
\alias{print.beastList}
\title{print}
\usage{
\method{print}{beastList}(x, ...)
}
\arguments{
\item{x}{a list of beast object}

\item{...}{no used}
}
\value{
message
}
\description{
print information of a list of beast trees
}
\author{
Guangchuang Yu
}
