% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{r8s-class}
\alias{r8s-class}
\title{Class "r8s"
This class stores output info from r8s}
\description{
Class "r8s"
This class stores output info from r8s
}
\section{Slots}{

\describe{
\item{\code{file}}{input file}

\item{\code{fields}}{available feature}

\item{\code{treetext}}{tree text}

\item{\code{phylo}}{multiPhylo, time tree, rate tree and absolute substitution tree}

\item{\code{extraInfo}}{extra information}
}}
\author{
Guangchuang Yu \url{http://guangchuangyu.github.io}
}
\keyword{classes}

