% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{range_format}
\alias{range_format}
\title{range_format}
\usage{
range_format(x, trans = NULL)
}
\arguments{
\item{x}{input list}

\item{trans}{transformation function}
}
\value{
character vector of '[lower, upper]'
}
\description{
format a list of range (HPD, CI, etc that has length of 2)
}
\author{
Guangchuang Yu
}
