% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{raxml-class}
\alias{raxml-class}
\title{Class "raxml"
This class stores RAxML bootstrapping analysis result}
\description{
Class "raxml"
This class stores RAxML bootstrapping analysis result
}
\section{Slots}{

\describe{
\item{\code{file}}{input file}

\item{\code{fields}}{available features}

\item{\code{treetext}}{tree text}

\item{\code{phylo}}{phylo object of treetext}

\item{\code{bootstrap}}{bootstrap value}

\item{\code{extraInfo}}{extra information}
}}
\author{
Guangchuang Yu \url{http://ygc.name}
}
\keyword{classes}

