% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raxml2nwk.R
\name{raxml2nwk}
\alias{raxml2nwk}
\title{raxml2nwk}
\usage{
raxml2nwk(infile, outfile = "raxml.tree")
}
\arguments{
\item{infile}{input file}

\item{outfile}{output file}
}
\value{
newick file
}
\description{
convert raxml bootstrap tree to newick format
}
\author{
Guangchuang Yu
}

