% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paml_rst.R
\name{read.baseml}
\alias{read.baseml}
\title{read.baseml}
\usage{
read.baseml(rstfile, mlbfile)
}
\arguments{
\item{rstfile}{rst file}

\item{mlbfile}{mlb file}
}
\value{
A \code{paml_rst} object
}
\description{
read rst and mlb file from baseml output
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

