% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/beast.R
\name{read.beast}
\alias{read.beast}
\title{read.beast}
\usage{
read.beast(file)
}
\arguments{
\item{file}{beast file}
}
\value{
\code{beast} object
}
\description{
read beast output
}
\examples{
file <- system.file("extdata", "beast_mcc.tree", package="ggtree")
read.beast(file)
}
\author{
Yu Guangchuang
}

