% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/codeml.R
\name{read.codeml}
\alias{read.codeml}
\title{read.codeml}
\usage{
read.codeml(rstfile, mlcfile)
}
\arguments{
\item{rstfile}{rst file}

\item{mlcfile}{mlc file}
}
\value{
A \code{codeml} object
}
\description{
read baseml output
}
\author{
ygc
}

