% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paml.R
\name{read.codeml_mlc}
\alias{read.codeml_mlc}
\title{read.codeml_mlc}
\usage{
read.codeml_mlc(mlcfile)
}
\arguments{
\item{mlcfile}{mlc file}
}
\value{
A \code{codeml_mlc} object
}
\description{
read mlc file of codeml output
}
\author{
ygc
}

