% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hyphy.R
\name{read.hyphy}
\alias{read.hyphy}
\title{nwk <- system.file("extdata/HYPHY", "labelledtree.tree", package="ggtree")
ancseq <- system.file("extdata/HYPHY", "ancseq.nex", package="ggtree")}
\usage{
read.hyphy(nwk, ancseq, tip.fasfile = NULL)
}
\description{
nwk <- system.file("extdata/HYPHY", "labelledtree.tree", package="ggtree")
ancseq <- system.file("extdata/HYPHY", "ancseq.nex", package="ggtree")
}

