% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jplace.R
\name{read.jplace}
\alias{read.jplace}
\title{read.jplace}
\usage{
read.jplace(file)
}
\arguments{
\item{file}{jplace file}
}
\value{
\code{jplace} instance
}
\description{
read jplace file
}
\examples{
jp <- system.file("extdata", "sample.jplace", package="ggtree")
read.jplace(jp)
}
\author{
ygc
}

