% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paml_rst.R
\name{read.paml_rst}
\alias{read.paml_rst}
\title{read.paml_rst}
\usage{
read.paml_rst(rstfile, tip.fasfile = NULL)
}
\arguments{
\item{rstfile}{rst file}

\item{tip.fasfile}{fasta file of tips}
}
\value{
A \code{paml_rst} object
}
\description{
read rst file from paml output
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

