% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylip.R
\name{read.phylip}
\alias{read.phylip}
\title{read.phylip}
\usage{
read.phylip(file)
}
\arguments{
\item{file}{phylip file}
}
\value{
an instance of 'phylip'
}
\description{
parsing phylip tree format
}
\author{
Guangchuang Yu
}

