% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RAxML.R
\name{read.raxml}
\alias{read.raxml}
\title{read.raxml}
\usage{
read.raxml(file)
}
\arguments{
\item{file}{RAxML bootstrapping analysis output}
}
\value{
raxml object
}
\description{
parse RAxML bootstrapping analysis output
}
\author{
Guangchuang Yu
}

