% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ape.R
\name{read.tree}
\alias{read.tree}
\title{read newick tree}
\source{
This is just the imported function
from the ape package. The documentation you should
read for the read.tree function can be found here: \link[ape]{read.tree}
}
\usage{
read.tree(file = "", text = NULL, tree.names = NULL, skip = 0,
  comment.char = "#", keep.multi = FALSE, ...)
}
\arguments{
\item{file}{file name}

\item{text}{alternatively, using newick text}

\item{tree.names}{if read several trees, specify their names}

\item{skip}{number of lines of the input file to skip}

\item{comment.char}{a single character,
the remaining of the line after this character is ignored.}

\item{keep.multi}{if 'TRUE' and 'tree.names = NULL'
then single trees are returned in 'multiPhylo' format
with any name that is present. Default is 'FALSE'}

\item{...}{further arguments to be passed to 'scan()'.}
}
\description{
read newick tree
}
\seealso{
\link[ape]{read.tree}
}

