% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{\%<\%}
\alias{\%<\%}
\title{\%<\%}
\usage{
pg \%<\% tree
}
\arguments{
\item{pg}{ggplot2 object}

\item{tree}{phylogenetic tree}
}
\value{
updated ggplot object
}
\description{
update tree viewing with a new tree
}
\author{
Yu Guangchuang
}

