% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/method-reroot.R
\docType{methods}
\name{reroot}
\alias{reroot}
\alias{reroot,phylo-method}
\title{reroot method}
\usage{
reroot(object, node, ...)

\S4method{reroot}{phylo}(object, node, ...)
}
\arguments{
\item{object}{\code{treedata} object}

\item{node}{internal nnode number}

\item{...}{additional parameter}
}
\value{
tree object
}
\description{
reroot method

reroot a tree
}
