% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_tree.R
\name{rescale_tree}
\alias{rescale_tree}
\title{rescale_tree}
\usage{
rescale_tree(tree_object, branch.length)
}
\arguments{
\item{tree_object}{tree object}

\item{branch.length}{numerical features (e.g. dN/dS)}
}
\value{
update tree object
}
\description{
rescale branch length of tree object
}
\author{
Guangchuang Yu
}
