% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{revts}
\alias{revts}
\title{revts}
\usage{
revts(treeview)
}
\arguments{
\item{treeview}{original tree view}
}
\value{
updated tree view
}
\description{
reverse timescle x-axis by setting the most recent tip to 0
}
\details{
'scale_x_continuous(labels=abs)' is required if users want to set the x-axis lable to absolute value
}
\examples{
tr <- rtree(10)
p <- ggtree(tr) + theme_tree2()
p2 <- revts(p) 
p2 + scale_x_continuous(labels=abs)
}
\author{
Guangchuang Yu
}
