% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/treeIO.R
\name{rm.singleton.newick}
\alias{rm.singleton.newick}
\title{rm.singleton.newick}
\usage{
rm.singleton.newick(nwk, outfile = NULL)
}
\arguments{
\item{nwk}{newick file}

\item{outfile}{output newick file}
}
\value{
tree text
}
\description{
remove singleton
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

