% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytree.R
\name{rotateTreePoints.df}
\alias{rotateTreePoints.df}
\title{rotateTreePoints.data.fram}
\usage{
rotateTreePoints.df(df, pivot_node, nodes, angle)
}
\arguments{
\item{df}{tree data.frame}

\item{pivot_node}{is the id of the pivot node.}

\item{nodes}{list of node numbers that are to be rotated by angle around the pivot_node}

\item{angle}{in range [0,2], ie degrees/180, radians/pi}
}
\value{
updated tree data.frame with points rotated by angle
}
\description{
Rotate the points in a tree data.frame around a pivot node by the angle specified.
}

