% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ape.R
\name{rtree}
\alias{rtree}
\title{generate random tree}
\source{
This is just the imported function
from the ape package. The documentation you should
read for the rtree function can be found here: \link[ape]{rtree}
}
\usage{
rtree(n, rooted = TRUE, tip.label = NULL, br = runif, ...)
}
\arguments{
\item{n}{number of tips in the tree}

\item{rooted}{logcial}

\item{tip.label}{tip label}

\item{br}{one of the following: (i) an R function used to generate the
branch lengths ('rtree'; use 'NULL' to simulate only a
topology), or the coalescence times ('rcoal'); (ii) a
character to simulate a genuine coalescent tree for 'rcoal'
(the default); or (iii) a numeric vector for the branch
lengths or the coalescence times.}

\item{...}{additional parameters to be passed to 'br'}
}
\description{
generate random tree
}
\seealso{
\link[ape]{rtree}
}

