% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/RAxML.R, R/ape.R, R/codeml.R, R/codeml_mlc.R, R/hyphy.R, R/jplace.R, R/method-scale-color.R, R/phangorn.R, R/phylo.R, R/r8s.R
\docType{methods}
\name{scale_color}
\alias{scale_color}
\alias{scale_color,apeBootstrap-method}
\alias{scale_color,beast-method}
\alias{scale_color,codeml-method}
\alias{scale_color,codeml_mlc-method}
\alias{scale_color,hyphy-method}
\alias{scale_color,jplace-method}
\alias{scale_color,nhx-method}
\alias{scale_color,paml_rst-method}
\alias{scale_color,phangorn-method}
\alias{scale_color,phylo-method}
\alias{scale_color,r8s-method}
\alias{scale_color,raxml-method}
\title{scale_color method}
\usage{
scale_color(object, by, ...)

\S4method{scale_color}{raxml}(object, by = "bootstrap", ...)

\S4method{scale_color}{apeBootstrap}(object, by = "bootstrap", ...)

\S4method{scale_color}{codeml}(object, by, ...)

\S4method{scale_color}{codeml_mlc}(object, by, ...)

\S4method{scale_color}{hyphy}(object, by, ...)

\S4method{scale_color}{jplace}(object, by, ...)

\S4method{scale_color}{beast}(object, by, ...)

\S4method{scale_color}{nhx}(object, by, ...)

\S4method{scale_color}{paml_rst}(object, by, ...)

\S4method{scale_color}{phangorn}(object, by, ...)

\S4method{scale_color}{phylo}(object, by, ...)

\S4method{scale_color}{r8s}(object, by = "bootstrap", tree = "TREE")
}
\arguments{
\item{object}{supported objects, including phylo, paml_rst,
codeml_mlc, codeml, jplace, beast, hyphy}

\item{by}{one of numerical attributes}

\item{...}{additional parameter}

\item{tree}{which tree selected}
}
\value{
color vector
}
\description{
scale color by a numerical tree attribute
}

