% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_x_ggtree}
\alias{scale_x_ggtree}
\title{scale_x_ggtree}
\usage{
scale_x_ggtree(breaks = waiver(), labels = waiver())
}
\arguments{
\item{breaks}{breaks for tree}

\item{labels}{lables for corresponding breaks}
}
\value{
updated tree view
}
\description{
scale x for tree with gheatmap
}
\author{
Guangchuang Yu
}
