% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gplot.R
\name{scale_x_heatmap}
\alias{scale_x_heatmap}
\title{scale_x_heatmap}
\usage{
scale_x_heatmap(p, breaks, labels = NULL)
}
\arguments{
\item{p}{tree view}

\item{breaks}{breaks for tree}

\item{labels}{lables for corresponding breaks}
}
\value{
tree view
}
\description{
scale x for tree with heatmap
}
\author{
Guangchuang Yu
}

