% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x-axis.R
\name{scale_x_range}
\alias{scale_x_range}
\title{scale_x_range}
\usage{
scale_x_range(center = "x")
}
\arguments{
\item{center}{center of the range}
}
\value{
ggtree object
}
\description{
second x-axis for geom_range
}
\author{
Guangchuang Yu
}
