% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{scaleClade}
\alias{scaleClade}
\title{scaleClade}
\usage{
scaleClade(tree_view, node, scale = 1)
}
\arguments{
\item{tree_view}{tree view}

\item{node}{clade node}

\item{scale}{scale}
}
\value{
tree view
}
\description{
scale clade
}
\author{
Guangchuang Yu
}

