% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/paml_rst.R
\docType{methods}
\name{"set.subs<-"}
\alias{"set.subs<-"}
\alias{set.subs<-}
\alias{set.subs<-,paml_rst-method,ANY-method}
\title{set.subs<- method}
\usage{
set.subs(x, type) <- value

set.subs(x, type) <- value
}
\arguments{
\item{x}{one of \code{paml_rst}, \code{codeml} object}

\item{type}{type}

\item{value}{subs}
}
\value{
updated \code{paml_rst}, \code{codeml} object
}
\description{
set.subs<- method
}

