% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_plot.R
\name{set_tree_xlim}
\alias{set_tree_xlim}
\title{set_tree_xlim}
\usage{
set_tree_xlim(tree_view, xlim)
}
\arguments{
\item{tree_view}{tree view}

\item{xlim}{xlim, should be of length 2}
}
\value{
updated tree view
}
\description{
set x axis limits for Tree panel
}
\author{
guangchuang yu
}

