% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/beast.R, R/jplace.R
\docType{methods}
\name{show,beast-method}
\alias{show}
\alias{show,beast-method}
\title{show method}
\usage{
\S4method{show}{beast}(object)

show(object)
}
\arguments{
\item{object}{one of \code{jplace}, \code{beast} object}
}
\value{
print info
}
\description{
show method for \code{jplace} instance
}
\examples{
jp <- system.file("extdata", "sample.jplace", package="ggtree")
jp <- read.jplace(jp)
show(jp)
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

