% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_balance.R
\name{stat_balance}
\alias{stat_balance}
\title{stat_balance}
\usage{
stat_balance(mapping = NULL, data = NULL, geom = "rect",
  position = "identity", node, show.legend = NA, inherit.aes = FALSE,
  fill, color, alpha, extend = 0, extendto = NULL, ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{geom}{geometric object}

\item{position}{position}

\item{node}{node number}

\item{show.legend}{show legend}

\item{inherit.aes}{logical}

\item{fill}{fill color}

\item{color}{color to outline highlights and divide balance}

\item{alpha}{transparency}

\item{extend}{extend xmax of the rectangle}

\item{extendto}{extend xmax to extendto}

\item{...}{additional parameter}
}
\value{
layer
}
\description{
stat_balance
}
