% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_chull.R
\name{stat_chull}
\alias{stat_chull}
\title{stat_chull}
\usage{
stat_chull(mapping = NULL, data = NULL, geom = "polygon",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\value{
ggplot2
}
\description{
layer of hilight clade with rectangle
}
