% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hilight.R
\name{stat_hilight}
\alias{stat_hilight}
\title{stat_hilight}
\usage{
stat_hilight(mapping = NULL, data = NULL, geom = "rect",
  position = "identity", node, show.legend = NA, inherit.aes = FALSE,
  fill, alpha, extend = 0, xmax = NULL, ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{geom}{geometric object}

\item{position}{position}

\item{node}{node number}

\item{show.legend}{show legend}

\item{inherit.aes}{logical}

\item{fill}{fill color}

\item{alpha}{transparency}

\item{extend}{extend xmax of the rectangle}

\item{...}{additional parameter}

\item{extendto}{extend xmax to extendto}
}
\value{
layer
}
\description{
stat_hilight
}

