% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subview.R
\name{subview}
\alias{subview}
\title{subview}
\usage{
subview(mainview, subview, x, y, width = 0.1, height = 0.1)
}
\arguments{
\item{mainview}{main view}

\item{subview}{sub view}

\item{x}{x position}

\item{y}{y position}

\item{width}{width of subview, [0,1]}

\item{height}{height of subview, [0,1]}
}
\value{
ggplot object
}
\description{
add subview to mainview for ggplot2 objects
}
\author{
Guangchuang Yu
}

