% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treedata-function.R
\name{td_mutate}
\alias{td_mutate}
\title{td-mutate}
\usage{
td_mutate(..., .f = NULL)
}
\arguments{
\item{...}{additional parameters that pass to dplyr::mutate}

\item{.f}{a function (if any, defaults to NULL) that pre-operate the data}
}
\value{
A function to mutate ggtree plot data
}
\description{
mutate data for tree annotation layer
}
\details{
The 'td_mutate()' function returns another function that can be
used to mutate ggtree() plot data. The function can be passed to the 'data' parameter
of geom layer to perform adding new variables and preserving existing ones.
}
\seealso{
\link[dplyr:mutate]{mutate}
}
