% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treedata-function.R
\name{td_unnest}
\alias{td_unnest}
\title{td-unnest}
\usage{
td_unnest(cols, ..., .f = NULL)
}
\arguments{
\item{cols}{columns to unnest}

\item{...}{additional parameters that pass to tidyr::unnest}

\item{.f}{a function (if any, default is NULL) that pre-operate the data}
}
\value{
A function to unnest ggtree plot data
}
\description{
flatterns a list-column of data frame
}
\details{
The 'td_unnest' function returns another function that can be
used to unnest ggtree() plot data. The function can be passed to
the 'data' parameter of a geom layer to flattern list-cloumn tree data.
}
\author{
Guangchuang Yu
}
