% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_dendrogram}
\alias{theme_dendrogram}
\title{theme_dendrogram}
\usage{
theme_dendrogram(bgcolor = "white", fgcolor = "black", ...)
}
\arguments{
\item{bgcolor}{background color}

\item{fgcolor}{color for axis}

\item{...}{additional parameter}
}
\description{
dendrogram theme
}
\author{
Guangchuang Yu
}
