% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_inset}
\alias{theme_inset}
\title{theme_inset}
\usage{
theme_inset(legend.position = "none", ...)
}
\arguments{
\item{legend.position}{position of legend}

\item{...}{additional parameter}
}
\value{
ggplot object
}
\description{
inset theme
}
\details{
theme for inset function
}
\author{
Guangchuang Yu
}
