% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{theme_tree}
\alias{theme_tree}
\title{theme_tree}
\usage{
theme_tree()
}
\value{
updated ggplot object with new theme
}
\description{
tree theme
}
\examples{
require(ape)
tr <- rtree(10)
ggtree(tr) + theme_tree()
}
\author{
Yu Guangchuang
}

