% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_tree2}
\alias{theme_tree2}
\title{theme_tree2}
\usage{
theme_tree2(bgcolor = "white", fgcolor = "black", ...)
}
\arguments{
\item{bgcolor}{set background color, defaults to "white"}

\item{fgcolor}{set foreground color, defaults to "black"}

\item{...}{additional parameter}
}
\value{
updated ggplot object with new theme
}
\description{
tree2 theme
}
\details{
'theme_tree2' supports displaying phylogenetic distance by setting x-axis
}
\examples{
require(ape)
tr <- rtree(10)
ggtree(tr) + theme_tree2()
}
\author{
Guangchuang Yu
}
