% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_rectangular}
\alias{layout_rectangular}
\alias{layout_circular}
\alias{layout_inward_circular}
\alias{layout_fan}
\alias{layout_dendrogram}
\title{layout_rectangular}
\usage{
layout_rectangular()

layout_circular()

layout_inward_circular(xlim = NULL)

layout_fan(angle = 180)

layout_dendrogram()
}
\arguments{
\item{xlim}{x limits}

\item{angle}{open tree at specific angle}
}
\description{
tree layout
}
\author{
Guangchuang Yu
}
