% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_function.R
\name{tree_filter}
\alias{tree_filter}
\title{tree_filter}
\usage{
tree_filter(...)
}
\arguments{
\item{...}{Expressions that return a logical value.}
}
\value{
A function to filter ggtree plot data using conditions defined by '...'.
}
\description{
filter data for tree annotation layer
}
\details{
The 'tree_filter()' function returns another function that can be
used to subset ggtree() plot data. The function can be passed to the 'data' parameter
in geom layer to perform subsetting. All rows that satisy your conditions will be retained.
}
\examples{
tree <- rtree(30)
## similar to 'ggtree(tree) + geom_tippoint()'
ggtree(tree) + geom_point(data = tree_filter(isTip))
}
\seealso{
\link[dplyr:filter]{filter}
}
\author{
Guangchuang Yu
}
