% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{\%<\%}
\alias{\%<\%}
\title{\%<\%}
\usage{
pg \%<\% x
}
\arguments{
\item{pg}{ggplot2 object}

\item{x}{update by x}
}
\value{
updated ggplot object
}
\description{
update tree
}
\examples{
library("ggplot2")
nwk <- system.file("extdata", "sample.nwk", package="ggtree")
tree <- read.tree(nwk)
p <- ggtree(tree) + geom_tippoint(color="#b5e521", alpha=1/4, size=10)
p \%<\% rtree(30)
}
\author{
Yu Guangchuang
}
