% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{\%<\%}
\alias{\%<\%}
\title{\%<\%}
\usage{
pg \%<\% x
}
\arguments{
\item{pg}{ggplot2 object}

\item{x}{update by x}
}
\value{
updated ggplot object
}
\description{
update tree
}
\author{
Yu Guangchuang
}

