% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jplace.R
\name{write.jplace}
\alias{write.jplace}
\title{write.jplace}
\usage{
write.jplace(nwk, data, outfile)
}
\arguments{
\item{nwk}{tree in newick format}

\item{data}{annotation data}

\item{outfile}{jplace output file}
}
\value{
jplace file
}
\description{
generate jplace file
}
\examples{
tree <- system.file("extdata", "pa.nwk", package="ggtree")
data <- read.csv(system.file("extdata", "pa_subs.csv", package="ggtree"),
                stringsAsFactor=FALSE)
outfile <- tempfile()
write.jplace(tree, data, outfile)
}
\author{
ygc
}

