% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x-axis.R
\name{ylim_ggtree}
\alias{ylim_ggtree}
\title{ylim_ggtree}
\usage{
ylim_ggtree(ggtree, expand_limits = expand_scale(0, 0.6))
}
\arguments{
\item{ggtree}{ggtree object}

\item{expand_limits}{amount to expand the limits}
}
\value{
ggplot2 object with new y limits
}
\description{
Set y axis limits of a ggplot based on the y limits of a ggtree.
This is useful for using cowplot or patchwork to properly align ggtree with other ggplot objects.
}
\author{
Guangchuang Yu
}
