% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade-functions.R
\name{zoomClade}
\alias{zoomClade}
\title{zoomClade}
\usage{
zoomClade(tree_view = NULL, node)
}
\arguments{
\item{tree_view}{tree view (i.e. the ggtree object). If tree_view is NULL, the last ggplot will be used.}

\item{node}{internal node number}
}
\value{
full tree with zoom in clade
}
\description{
zoom selected clade of a tree
}
\author{
Guangchuang Yu
}
