% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree.R
\name{\%<+\%}
\alias{\%<+\%}
\title{\%<+\%}
\usage{
pg \%<+\% data
}
\arguments{
\item{pg}{ggplot2 object}

\item{data}{annotation data}
}
\value{
ggplot object with annotation data added
}
\description{
add annotation data to a tree
}
\examples{
nwk <- system.file("extdata", "sample.nwk", package="ggtree")
tree <- read.tree(nwk)
p <- ggtree(tree)
dd <- data.frame(taxa=LETTERS[1:13],
   		 place=c(rep("GZ", 5), rep("HK", 3), rep("CZ", 4), NA),
             value=round(abs(rnorm(13, mean=70, sd=10)), digits=1))
row.names(dd) <- NULL
p \%<+\% dd + geom_text(aes(color=place, label=label), hjust=-0.5)
}
\author{
Yu Guangchuang
}

