% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{\%+>\%}
\alias{\%+>\%}
\title{\%+>\%}
\usage{
p \%+>\% data
}
\arguments{
\item{p}{tree view}

\item{data}{data.frame}
}
\value{
updated data.frame
}
\description{
update data with tree info (y coordination and panel)
}
\author{
Guangchuang Yu
}
