% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{xlim_align}
\alias{xlim_align}
\alias{ylim_align}
\title{xlim_align}
\usage{
xlim_align(gg, limits = NULL)

ylim_align(gg, limits = NULL)
}
\arguments{
\item{gg}{ggplot object}

\item{limits}{vector of limits. If NULL, determine from \code{gg}.}
}
\value{
ggplot2 object with new limits
}
\description{
set axis limits (x or y) of a \code{ggplot} object (left hand side of \code{+})
based on the x (\code{xlim_align}) or y (\code{ylim_align}) limits of another \code{ggplot} object (right hand side of \code{+}).
This is useful for using \code{cowplot} or \code{patchwork} to align \code{ggplot} objects.
}
\author{
Guangchuang Yu
}
