% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{xlim2}
\alias{xlim2}
\alias{ylim2}
\title{xlim2}
\usage{
xlim2(gg, limits = NULL)

ylim2(gg, limits = NULL)
}
\arguments{
\item{gg}{ggplot object}

\item{limits}{vector of limits. If NULL, determine from \code{gg}.}
}
\value{
ggplot2 object with new limits
}
\description{
set axis limits (x or y) of a \code{ggplot} object (left hand side of \code{+})
based on the x (\code{xlim2}) or y (\code{ylim2}) limits of another \code{ggplot} object (right hand side of \code{+}).
This is useful for using \code{cowplot} or \code{patchwork} to align \code{ggplot} objects.
}
\author{
Guangchuang Yu
}
