% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape.R
\name{apeBoot}
\alias{apeBoot}
\title{apeBoot}
\usage{
apeBoot(phylo, boot)
}
\arguments{
\item{phylo}{phylo}

\item{boot}{bootstrap values}
}
\value{
an instance of 'apeBootstrap'
}
\description{
merge phylo and output of boot.phylo to 'apeBootstrap' object
}
\author{
Guangchuang Yu
}
