% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytree.R
\name{applyLayoutDaylight}
\alias{applyLayoutDaylight}
\title{applyLayoutDaylight}
\usage{
applyLayoutDaylight(df, node_id)
}
\arguments{
\item{df}{tree data.frame}

\item{node_id}{is id of the node from which daylight is measured to the other subtrees.}
}
\value{
list with tree data.frame with updated layout using daylight algorithm and max_change angle.
}
\description{
Apply the daylight alorithm to adjust the spacing between the subtrees and tips of the
specified node.
}

