% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-fortify.R
\name{as.data.frame.phylo}
\alias{as.data.frame.phylo}
\title{as.data.frame}
\usage{
\method{as.data.frame}{phylo}(x, row.names, optional, layout = "rectangular",
  ...)
}
\arguments{
\item{x}{phylo object}

\item{row.names}{omitted here}

\item{optional}{omitted here}

\item{layout}{layout}

\item{...}{additional parameter}
}
\value{
data.frame
}
\description{
convert phylo to data.frame
}
\author{
Yu Guangchuang
}

