% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{beast-class}
\alias{beast-class}
\alias{get.tree,beast-method}
\title{Class "beast"
This class stores information of beast output}
\description{
Class "beast"
This class stores information of beast output
}
\section{Slots}{

\describe{
\item{\code{fields}}{beast statistic variables}

\item{\code{treetext}}{tree text in beast file}

\item{\code{phylo}}{tree phylo object}

\item{\code{translation}}{tip number to name translation in beast file}

\item{\code{stats}}{beast statistics}

\item{\code{file}}{beast file, nexus format}

\item{\code{extraInfo}}{extra information}
}}
\author{
Guangchuang Yu \url{http://guangchuangyu.github.io}
}
\seealso{
\code{\link{show}} \code{\link{get.fields}}
          \code{\link{ggtree}}
}
\keyword{classes}

