% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade-functions.R
\name{collapse.ggtree}
\alias{collapse.ggtree}
\title{collapse-ggtree}
\usage{
\method{collapse}{ggtree}(x = NULL, node, mode = "none", clade_name = NULL, ...)
}
\arguments{
\item{x}{tree view (i.e. the ggtree object). If tree_view is NULL, the last ggplot will be used.}

\item{node}{internal node number}

\item{mode}{one of 'none', 'max', 'min' and 'mixed'}

\item{clade_name}{set clade name. If clade_name = NULL, do nothing}

\item{...}{additional parameters}
}
\value{
tree view
}
\description{
collapse a clade
}
\seealso{
expand
}
\author{
Guangchuang Yu
}
