% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade-functions.R
\name{collapse}
\alias{collapse}
\title{collapse}
\usage{
collapse(tree_view = NULL, node)
}
\arguments{
\item{tree_view}{tree view}

\item{node}{clade node}
}
\value{
tree view
}
\description{
collapse a clade
}
\seealso{
expand
}
\author{
Guangchuang Yu
}
