% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{collapse}
\alias{collapse}
\title{collapse}
\usage{
collapse(tree_view, node)
}
\arguments{
\item{tree_view}{tree view}

\item{node}{clade node}
}
\value{
tree view
}
\description{
collapse a clade
}
\author{
Guangchuang Yu
}
\seealso{
expand
}

