% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-drop-tip.R
\docType{methods}
\name{drop.tip}
\alias{drop.tip}
\alias{drop.tip,nhx}
\title{drop.tip method}
\usage{
drop.tip(object, tip...)
}
\description{
drop.tip method
}
\author{
Casey Dunn \url{http://dunnlab.org}
}

