% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade-functions.R
\name{expand}
\alias{expand}
\title{expand}
\usage{
expand(tree_view = NULL, node)
}
\arguments{
\item{tree_view}{tree view (i.e. the ggtree object). If tree_view is NULL, the last ggplot will be used.}

\item{node}{internal node number to specify a clade. If NULL, using the whole tree}
}
\value{
tree view
}
\description{
expand collased clade
}
\seealso{
collapse
}
\author{
Guangchuang Yu
}
