% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{expand}
\alias{expand}
\title{expand}
\usage{
expand(tree_view, node)
}
\arguments{
\item{tree_view}{tree view}

\item{node}{clade node}
}
\value{
tree view
}
\description{
expand collased clade
}
\author{
Guangchuang Yu
}
\seealso{
collapse
}

