% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_plot.R
\name{facet_plot}
\alias{facet_plot}
\alias{geom_facet}
\title{facet_plot}
\usage{
facet_plot(p, panel, data, geom, mapping = NULL, ...)

geom_facet(panel, data, geom, mapping = NULL, ...)
}
\arguments{
\item{p}{tree view}

\item{panel}{panel name for plot of input data}

\item{data}{data to plot by 'geom', first column should be matched with tip label of tree}

\item{geom}{geom function to plot the data}

\item{mapping}{aes mapping for 'geom'}

\item{...}{additional parameters for 'geom'}
}
\value{
ggplot object
}
\description{
plot tree associated data in an additional panel
}
\details{
'facet_plot()' automatically re-arranges the input 'data' according to the tree structure,
visualizes the 'data' on specific 'panel' using the 'geom' function with aesthetic 'mapping' and other parameters,
and align the graph with the tree 'p' side by side. 'geom_facet' is a 'ggplot2' layer version of 'facet_plot'
}
\examples{
tr <- rtree(10)
dd = data.frame(id=tr$tip.label, value=abs(rnorm(10)))
p <- ggtree(tr)
facet_plot(p, 'Trait', data = dd, geom=geom_point, mapping=aes(x=value))
}
\author{
Guangchuang Yu
}
