% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_plot.R
\name{facet_data}
\alias{facet_data}
\title{facet_data}
\usage{
facet_data(tree_view, panel)
}
\arguments{
\item{tree_view}{ggtree object}

\item{panel}{data plotted in specific panel. If only one dataset used in the panel, return the data frame, else return a list of data frames.}
}
\value{
data frame or a list of data frames
}
\description{
extract data used in `facet_plot` or `geom_facet`
}
\author{
Guangchuang Yu
}
