% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade-functions.R
\name{flip}
\alias{flip}
\title{flip}
\usage{
flip(tree_view = NULL, node1, node2)
}
\arguments{
\item{tree_view}{tree view (i.e. the ggtree object). If tree_view is NULL, the last ggplot will be used.}

\item{node1}{node number of branch 1}

\item{node2}{node number of branch 2}
}
\value{
ggplot2 object
}
\description{
flip position of two selected branches
}
\author{
Guangchuang Yu
}
