% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{flip}
\alias{flip}
\title{flip}
\usage{
flip(tree_view, node1, node2)
}
\arguments{
\item{tree_view}{tree view}

\item{node1}{node number of branch 1}

\item{node2}{node number of branch 2}
}
\value{
ggplot2 object
}
\description{
flip position of two selected branches
}
\author{
Guangchuang Yu
}

