% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/treeIO.R
\name{fortify.phylo}
\alias{fortify.phylo}
\title{fortify}
\usage{
\method{fortify}{phylo}(model, data, ladderize = TRUE, right = FALSE, ...)
}
\arguments{
\item{model}{phylo object}

\item{data}{not use here}

\item{ladderize}{ladderize, logical}

\item{right}{logical}

\item{...}{additional parameter}
}
\value{
data.frame
}
\description{
fortify
}
\author{
Yu Guangchuang
}

