% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-fortify.R
\name{fortify.phylo}
\alias{fortify.phylo}
\title{fortify}
\usage{
\method{fortify}{phylo}(model, data, layout = "rectangular",
  ladderize = TRUE, right = FALSE, mrsd = NULL, as.Date = FALSE, ...)
}
\arguments{
\item{model}{phylo object}

\item{data}{not use here}

\item{layout}{layout}

\item{ladderize}{ladderize, logical}

\item{right}{logical}

\item{mrsd}{most recent sampling date}

\item{as.Date}{logical whether using Date class in time tree}

\item{...}{additional parameter}
}
\value{
data.frame
}
\description{
fortify a phylo to data.frame
}
\author{
Yu Guangchuang
}

