% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{geom_aline}
\alias{geom_aline}
\title{geom_aline}
\usage{
geom_aline(linetype = "dashed", ...)
}
\arguments{
\item{linetype}{line type}

\item{...}{additional parameter}
}
\value{
aline layer
}
\description{
add horizontal align lines
}
\examples{
require(ape)
tr <- rtree(10)
ggtree(tr) + geom_tiplab(align=TRUE) + geom_aline()
}
\author{
Yu Guangchuang
}

