% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_segment.R
\name{geom_aline}
\alias{geom_aline}
\title{geom_aline}
\usage{
geom_aline(mapping = NULL, linetype = "dotted", size = 1, ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{linetype}{line type}

\item{size}{line size}

\item{...}{additional parameter}
}
\value{
aline layer
}
\description{
add horizontal align lines
}
\author{
Yu Guangchuang
}

