% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_balance.R
\name{geom_balance}
\alias{geom_balance}
\title{geom_balance}
\usage{
geom_balance(node, fill = "steelblue", color = "white", alpha = 0.5,
  extend = 0, extendto = NULL)
}
\arguments{
\item{node}{selected node (balance) to hilight}

\item{fill}{color fill}

\item{color}{color to outline hilights and divide balance}

\item{alpha}{alpha (transparency)}

\item{extend}{extend xmax of the rectangle}

\item{extendto}{extend xmax to extendto}
}
\value{
ggplot2
}
\description{
hilight the two subclades of an internal node
}
\details{
Particularly useful when studying neighboring clades. Note that balances that
correspond to multichotomies will not be displayed.
}
\author{
Justin Silverman
}

