% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cladelabel.R
\name{geom_cladelabel}
\alias{geom_cladelabel}
\title{geom_cladelabel}
\usage{
geom_cladelabel(node, label, offset = 0, offset.text = 0, align = FALSE,
  barsize = 0.5, fontsize = 3.88, angle = 0, geom = "text", hjust = 0,
  fill = NA, ...)
}
\arguments{
\item{node}{selected node}

\item{label}{clade label}

\item{offset}{offset of bar and text from the clade}

\item{offset.text}{offset of text from bar}

\item{align}{logical}

\item{barsize}{size of bar}

\item{fontsize}{size of text}

\item{angle}{angle of text}

\item{geom}{one of 'text' or 'label'}

\item{hjust}{hjust}

\item{fill}{fill label background, only work with geom='label'}

\item{...}{additional parameter}
}
\value{
ggplot layers
}
\description{
annotate a clade with bar and text label
}
\author{
Guangchuang Yu
}

