% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hilight.R
\name{geom_hilight}
\alias{geom_hilight}
\title{geom_hilight}
\usage{
geom_hilight(node, fill = "steelblue", alpha = 0.5, extend = 0,
  extendto = NULL)
}
\arguments{
\item{node}{selected node to hilight (required)}

\item{fill}{color fill (default = steelblue)}

\item{alpha}{alpha transparency, (default = 0.5)}

\item{extend}{extend xmax of the rectangle (default = 0)}

\item{extendto}{extend xmax to extendto (default = NULL)}
}
\value{
ggplot2
}
\description{
layer of hilight clade with rectangle
}
\author{
Guangchuang Yu
}
