% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{geom_hilight}
\alias{geom_hilight}
\title{geom_hilight}
\usage{
geom_hilight(tree_object, node, ...)
}
\arguments{
\item{tree_object}{supported tree object}

\item{node}{internal node}

\item{...}{additional parameters}
}
\value{
ggplot layer
}
\description{
hilight clade with rectangle
}
\author{
Guangchuang Yu
}

