% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hilight.R
\name{geom_hilight}
\alias{geom_hilight}
\title{geom_hilight}
\usage{
geom_hilight(node, fill = "steelblue", alpha = 0.5, extend = 0,
  extendto = NULL)
}
\arguments{
\item{node}{selected node to hilight}

\item{fill}{color fill}

\item{alpha}{alpha (transparency)}

\item{extend}{extend xmax of the rectangle}

\item{extendto}{extend xmax to extendto}
}
\value{
ggplot2
}
\description{
layer of hilight clade with rectangle
}
\author{
Guangchuang Yu
}
