% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hilight_encircle.R
\name{geom_hilight_encircle}
\alias{geom_hilight_encircle}
\title{geom_hilight_encircle}
\usage{
geom_hilight_encircle(
  data = NULL,
  node,
  mapping = NULL,
  fill = "steelblue",
  alpha = 0.5,
  extend = 0,
  ...
)
}
\arguments{
\item{data}{data frame to calculate xspline (default = NULL)}

\item{node}{selected node to hilight (required)}

\item{mapping}{aesthetic mapping (default = NULL)}

\item{fill}{colour fill (default = steelblue)}

\item{alpha}{alpha (transparency) (default = 0.5)}

\item{extend}{expands the xspline clade region only (default = 0)}

\item{...}{addtional parameters, including:
'spread' spread of shape? (default = 0.1),
'linetype' Line type of xspline (default = 1),
'size' Size of xspline line (default = 1),
's_shape' Corresponds to shape of xspline (default = 0.5),
's_open' Boolean switch determines if xspline shape is open or closed. (default = FALSE)}
}
\value{
ggplot2
}
\description{
layer of hilight clade with xspline
}
