% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_label.R
\name{geom_label2}
\alias{geom_label2}
\title{geom_label2}
\usage{
geom_label2(mapping = NULL, data = NULL, ..., position = "identity",
  family = "sans", parse = FALSE, nudge_x = 0, nudge_y = 0,
  label.padding = unit(0.25, "lines"), label.r = unit(0.15, "lines"),
  label.size = 0.25, na.rm = TRUE, show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{the aesthetic mapping}

\item{data}{A layer specific dataset -
only needed if you want to override he plot defaults.}

\item{...}{other arguments passed on to 'layer'}

\item{position}{The position adjustment to use for overlapping points on this layer}

\item{family}{sans by default, can be any supported font}

\item{parse}{if TRUE, the labels will be passd into expressions}

\item{nudge_x}{horizontal adjustment}

\item{nudge_y}{vertical adjustment}

\item{label.padding}{Amount of padding around label.}

\item{label.r}{Radius of rounded corners.}

\item{label.size}{Size of label border, in mm}

\item{na.rm}{logical}

\item{show.legend}{logical}

\item{inherit.aes}{logical}
}
\value{
label layer
}
\description{
geom_label2 support aes(subset) via setup_data
}
\author{
Guangchuang Yu
}
\seealso{
\link[ggplot2]{geom_label}
}

