% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_nodelab.R
\name{geom_nodelab}
\alias{geom_nodelab}
\title{geom_nodelab}
\usage{
geom_nodelab(
  mapping = NULL,
  nudge_x = 0,
  nudge_y = 0,
  geom = "text",
  hjust = 0.5,
  ...
)
}
\arguments{
\item{mapping}{aes mapping}

\item{nudge_x}{horizontal adjustment to nudge label}

\item{nudge_y}{vertical adjustment to nudge label}

\item{geom}{one of 'text', "shadowtext", 'label', 'image' and 'phylopic'}

\item{hjust}{horizontal alignment, one of 0, 0.5 or 1}

\item{...}{additional parameters, see also
the additional parameters of \code{\link[=geom_tiplab]{geom_tiplab()}}.}
}
\value{
geom layer
}
\description{
add node label layer
}
\seealso{
\code{\link[=geom_tiplab]{geom_tiplab()}}
}
\author{
Guangchuang Yu
}
