% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{geom_nodelab2}
\alias{geom_nodelab2}
\title{geom_nodelab2}
\usage{
geom_nodelab2(
  mapping = NULL,
  nudge_x = 0,
  nudge_y = 0,
  geom = "text",
  hjust = 0.5,
  ...
)
}
\arguments{
\item{mapping}{aes mapping}

\item{nudge_x}{horizontal adjustment to nudge label}

\item{nudge_y}{vertical adjustment to nudge label}

\item{geom}{one of 'text', 'label', 'image' and 'phylopic'}

\item{hjust}{horizontal alignment, one of 0, 0.5 or 1}

\item{...}{additional parameters}
}
\value{
node label layer
}
\description{
add node label for circular layout
}
\author{
Guangchuang Yu
}
