% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_point.R
\name{geom_nodepoint}
\alias{geom_nodepoint}
\title{geom_nodepoint}
\usage{
geom_nodepoint(mapping = NULL, ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{...}{additional parameter}
}
\value{
node point layer
}
\description{
add node point
}
\author{
Guangchuang Yu
}

