% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point.R
\name{geom_nodepoint}
\alias{geom_nodepoint}
\title{geom_nodepoint}
\usage{
geom_nodepoint(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{position}{position}

\item{na.rm}{logical}

\item{show.legend}{logical}

\item{inherit.aes}{logical}

\item{...}{addktional parameter}
}
\value{
node point layer
}
\description{
add node point
}
\author{
Guangchuang Yu
}

