% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{geom_place}
\alias{geom_place}
\title{geom_place}
\usage{
geom_place(data, map, place, by = "node", ...)
}
\arguments{
\item{data}{data}

\item{map}{mapping variable}

\item{place}{placement info}

\item{by}{one of "node" and "edge"}

\item{...}{additional parameter}
}
\value{
text layer
}
\description{
place annotation in tree
}
\author{
ygc
}

