% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point.R
\name{geom_point2}
\alias{geom_point2}
\title{geom_point2}
\usage{
geom_point2(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{position}{position}

\item{na.rm}{logical}

\item{show.legend}{logical}

\item{inherit.aes}{logical}

\item{...}{addktional parameter}
}
\value{
point layer
}
\description{
geom_point2 support aes(subset) via setup_data
}
\author{
Guangchuang Yu
}
\seealso{
\link[ggplot2]{geom_point}
}

