% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_range.R
\name{geom_range}
\alias{geom_range}
\title{geom_range}
\usage{
geom_range(range, center = "auto", ...)
}
\arguments{
\item{range}{range(interval) to be displayed, e.g. "height_0.95_HPD"}

\item{center}{center of the range, mean, median or auto (default, the center of the range)}

\item{...}{additional parameter, e.g. color, size, alpha}
}
\value{
ggplot layer
}
\description{
horizontal bar of range (HPD, range etc) on nodes to present uncertainty of evolutionary inference
}
\references{
For demonstration of this function, please refer to chapter 5.2.4 of
\emph{Data Integration, Manipulation and Visualization of Phylogenetic Trees}
\url{http://yulab-smu.top/treedata-book/index.html} by Guangchuang Yu.
}
\author{
Guangchuang Yu
}
