% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_range.R
\name{geom_range}
\alias{geom_range}
\title{geom_range}
\usage{
geom_range(range = "length_0.95_HPD", center = "x", ...)
}
\arguments{
\item{range}{range, e.g. "height_0.95_HPD"}

\item{center}{center of the range, mean or median}

\item{...}{additional parameter, e.g. color, size, alpha}
}
\value{
ggplot layer
}
\description{
bar of range (HPD, range etc) to present uncertainty of evolutionary inference
}
\author{
Guangchuang Yu
}
