% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_rootedge.R
\name{geom_rootedge}
\alias{geom_rootedge}
\title{geom_rootedge}
\usage{
geom_rootedge(rootedge = NULL)
}
\arguments{
\item{rootedge}{length of rootedge; use phylo$root.edge if rootedge = NULL (by default).}
}
\value{
ggplot layer
}
\description{
display root edge
}
\author{
Guangchuang Yu
}
