% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point.R
\name{geom_rootpoint}
\alias{geom_rootpoint}
\title{geom_rootpoint}
\usage{
geom_rootpoint(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mapping created by \code{aes()}.
If \code{inherit.aes = TRUE}, the mapping can be inherited from the plot mapping as
specified in the call to \code{ggplot()}.}

\item{data}{The data to be displayed in this layer. If 'NULL' (the default),
the data is inherited from the plot data as specified in the call to 'ggplot()',}

\item{position}{Position adjustment.}

\item{na.rm}{logical. If 'FALSE' (the default), missing values are removed with a warning. If 'TRUE', missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
'NA', the default, includes if any aesthetics are mapped. 'FALSE' never includes, and 'TRUE' always includes.}

\item{inherit.aes}{logical (default is 'TRUE'). If 'FALSE', overrides the default aesthetics,
rather then combining with them.}

\item{...}{addtional parameters that passed on to this layer. These are often aesthetics, used to set an aesthetic to a fixed value, like \code{colour = "red"} or \code{size = 3}.}
}
\value{
root point layer
}
\description{
add root point
}
\author{
Guangchuang Yu
}
