% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_point.R
\name{geom_rootpoint}
\alias{geom_rootpoint}
\title{geom_rootpoint}
\usage{
geom_rootpoint(mapping = NULL, ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{...}{additional parameter}
}
\value{
root point layer
}
\description{
add root point
}
\author{
Guangchuang Yu
}

