% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point.R
\name{geom_rootpoint}
\alias{geom_rootpoint}
\title{geom_rootpoint}
\usage{
geom_rootpoint(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{position}{position}

\item{na.rm}{logical}

\item{show.legend}{logical}

\item{inherit.aes}{logical}

\item{...}{addktional parameter}
}
\value{
root point layer
}
\description{
add root point
}
\author{
Guangchuang Yu
}
