% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_segment.R
\name{geom_segment2}
\alias{geom_segment2}
\title{geom_segment2}
\usage{
geom_segment2(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", lineend = "butt", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, nudge_x = 0, arrow = NULL,
  arrow.fill = NULL, ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{stat}{Name of stat to modify data}

\item{position}{position}

\item{lineend}{lineend}

\item{na.rm}{logical}

\item{show.legend}{logical}

\item{inherit.aes}{logical}

\item{nudge_x}{horizontal adjustment of x}

\item{arrow}{specification for arrow heads, as created by arrow().}

\item{arrow.fill}{fill color to usse for the arrow head (if closed). `NULL` means use `colour` aesthetic.}

\item{...}{additional parameter}
}
\value{
add segment layer
}
\description{
geom_segment2 support aes(subset) via setup_data
}
\seealso{
\link[ggplot2]{geom_segment}
}
\author{
Guangchuang Yu
}
