% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_segment.R
\name{geom_segment2}
\alias{geom_segment2}
\title{geom_segment2}
\usage{
geom_segment2(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", arrow = NULL, lineend = "butt", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{arrow}{arrow}

\item{lineend}{lineend}

\item{na.rm}{logical}

\item{show.legend}{logical}

\item{inherit.aes}{logical}

\item{...}{additional parameter}
}
\value{
add segment layer
}
\description{
geom_segment2 support aes(subset) via setup_data
}
\author{
Guangchuang Yu
}
\seealso{
\link[ggplot2]{geom_segment}
}

