% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_strip.R
\name{geom_strip}
\alias{geom_strip}
\title{geom_strip}
\usage{
geom_strip(taxa1, taxa2, label = NA, offset = 0, offset.text = 0,
  align = TRUE, barsize = 0.5, barextend = 0, fontsize = 3.88,
  angle = 0, geom = "text", hjust = 0, fill = NA, family = "sans",
  parse = FALSE, ...)
}
\arguments{
\item{taxa1}{taxa1}

\item{taxa2}{taxa2}

\item{label}{optional label}

\item{offset}{offset of bar and text from the clade}

\item{offset.text}{offset of text from bar}

\item{align}{logical}

\item{barsize}{size of bar}

\item{barextend}{extend bar vertically}

\item{fontsize}{size of text}

\item{angle}{angle of text}

\item{geom}{one of 'text' or 'label'}

\item{hjust}{hjust}

\item{fill}{fill label background, only work with geom='label'}

\item{family}{sans by default, can be any supported font}

\item{parse}{logical, whether parse label}

\item{...}{additional parameter}
}
\value{
ggplot layers
}
\description{
annotate associated taxa (from taxa1 to taxa2, can be Monophyletic, Polyphyletic or Paraphyletc Taxa) with bar and (optional) text label
}
\author{
Guangchuang Yu
}

