% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_taxalink.R
\name{geom_taxalink}
\alias{geom_taxalink}
\title{geom_taxalink}
\usage{
geom_taxalink(taxa1, taxa2, curvature = 0.5, arrow = NULL,
  arrow.fill = NULL, ...)
}
\arguments{
\item{taxa1}{taxa1, can be label or node number}

\item{taxa2}{taxa2, can be label or node number}

\item{curvature}{A numeric value giving the amount of curvature.
Negative values produce left-hand curves,
positive values produce right-hand curves, and zero produces a straight line.}

\item{arrow}{specification for arrow heads, as created by arrow().}

\item{arrow.fill}{fill color to usse for the arrow head (if closed). `NULL` means use `colour` aesthetic.}

\item{...}{additional parameter}
}
\value{
ggplot layer
}
\description{
link between taxa
}
\author{
Guangchuang Yu
}
