% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_taxalink.R
\name{geom_taxalink}
\alias{geom_taxalink}
\title{geom_taxalink}
\usage{
geom_taxalink(
  taxa1,
  taxa2,
  curvature = 0.5,
  arrow = NULL,
  arrow.fill = NULL,
  offset = NULL,
  hratio = 1,
  outward = TRUE,
  ...
)
}
\arguments{
\item{taxa1}{taxa1, can be label or node number}

\item{taxa2}{taxa2, can be label or node number}

\item{curvature}{A numeric value giving the amount of curvature.
Negative values produce left-hand curves,
positive values produce right-hand curves, and zero produces a straight line.}

\item{arrow}{specification for arrow heads, as created by arrow().}

\item{arrow.fill}{fill color to usse for the arrow head (if closed). \code{NULL} means use \code{colour} aesthetic.}

\item{offset}{numeric, control the shift of curve line (the ratio of axis value,
range is "(0-1)"), default is NULL.}

\item{hratio}{numeric, the height of curve line, default is 1.}

\item{outward}{logical, control the orientation of curve when the layout of tree is circular,
fan or other layout in polar coordinate, default is TRUE.}

\item{...}{additional parameter.}
}
\value{
ggplot layer
}
\description{
link between taxa
}
\author{
Guangchuang Yu
}
