% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_text.R
\name{geom_text}
\alias{geom_text}
\title{text annotations}
\source{
This is just the imported function
from the ggplot2 package. The documentation you should
read for the geom_text function can be found here: \link[ggplot2]{geom_text}
}
\usage{
geom_text(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", parse = FALSE, ...)
}
\arguments{
\item{mapping}{the aesthetic mapping}

\item{data}{A layer specific dataset -
only needed if you want to override he plot defaults.}

\item{stat}{The statistical transformation to use on the data for this layer}

\item{position}{The position adjustment to use for overlapping points on this layer}

\item{parse}{if TRUE, the labels will be passd into expressions}

\item{...}{other arguments passed on to 'layer'}
}
\description{
text annotations
}
\seealso{
\link[ggplot2]{geom_text}
}

