% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_text.R
\name{geom_text2}
\alias{geom_text2}
\title{geom_text2}
\usage{
geom_text2(mapping = NULL, data = NULL, ..., position = "identity",
  family = "sans", parse = FALSE, na.rm = TRUE, show.legend = NA,
  inherit.aes = TRUE, nudge_x = 0, nudge_y = 0, check_overlap = FALSE)
}
\arguments{
\item{mapping}{the aesthetic mapping}

\item{data}{A layer specific dataset -
only needed if you want to override he plot defaults.}

\item{...}{other arguments passed on to 'layer'}

\item{position}{The position adjustment to use for overlapping points on this layer}

\item{family}{sans by default, can be any supported font}

\item{parse}{if TRUE, the labels will be passd into expressions}

\item{na.rm}{logical}

\item{show.legend}{logical}

\item{inherit.aes}{logical}

\item{nudge_x}{horizontal adjustment}

\item{nudge_y}{vertical adjustment}

\item{check_overlap}{if TRUE, text that overlaps previous text in the same layer will not be plotted}
}
\value{
text layer
}
\description{
geom_text2 support aes(subset) via setup_data
}
\author{
Guangchuang Yu
}
\seealso{
\link[ggplot2]{geom_text}
}

