% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{geom_tiplab}
\alias{geom_tiplab}
\title{geom_tiplab}
\usage{
geom_tiplab(align = FALSE, ...)
}
\arguments{
\item{align}{align tip lab or not, logical}

\item{...}{additional parameter}
}
\value{
tip label layer
}
\description{
add tip label layer
}
\author{
Yu Guangchuang
}

