% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tiplab.R
\name{geom_tiplab}
\alias{geom_tiplab}
\title{geom_tiplab}
\usage{
geom_tiplab(
  mapping = NULL,
  hjust = 0,
  align = FALSE,
  linetype = "dotted",
  linesize = 0.5,
  geom = "text",
  offset = 0,
  as_ylab = FALSE,
  ...
)
}
\arguments{
\item{mapping}{aes mapping}

\item{hjust}{horizontal adjustment}

\item{align}{align tip lab or not, logical}

\item{linetype}{linetype for adding line if align = TRUE}

\item{linesize}{line size of line if align = TRUE}

\item{geom}{one of 'text', 'label', 'image' and 'phylopic'}

\item{offset}{tiplab offset}

\item{as_ylab}{display tip labels as y-axis label, only works for rectangular and dendrogram layouts}

\item{...}{additional parameter}
}
\value{
tip label layer
}
\description{
add tip label layer
}
\examples{
require(ape)
tr <- rtree(10)
ggtree(tr) + geom_tiplab()
}
\author{
Guangchuang Yu
}
