% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{geom_tiplab}
\alias{geom_tiplab}
\title{geom_tiplab}
\usage{
geom_tiplab(align = FALSE, hjust = 0, ...)
}
\arguments{
\item{align}{align tip lab or not, logical}

\item{hjust}{horizontal adjustment}

\item{...}{additional parameter}
}
\value{
tip label layer
}
\description{
add tip label layer
}
\examples{
require(ape)
tr <- rtree(10)
ggtree(tr) + geom_tiplab()
}
\author{
Yu Guangchuang
}

