% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_point.R
\name{geom_tippoint}
\alias{geom_tippoint}
\title{geom_tippoint}
\usage{
geom_tippoint(mapping = NULL, ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{...}{additional parameter}
}
\value{
tip point layer
}
\description{
add tip point
}
\author{
Guangchuang Yu
}

