% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point.R
\name{geom_tippoint}
\alias{geom_tippoint}
\title{geom_tippoint}
\usage{
geom_tippoint(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{position}{position}

\item{na.rm}{logical}

\item{show.legend}{logical}

\item{inherit.aes}{logical}

\item{...}{addktional parameter}
}
\value{
tip point layer
}
\description{
add tip point
}
\author{
Guangchuang Yu
}

