% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tree.R
\name{geom_tree}
\alias{geom_tree}
\title{geom_tree}
\usage{
geom_tree(
  mapping = NULL,
  data = NULL,
  layout = "rectangular",
  multiPhylo = FALSE,
  ...
)
}
\arguments{
\item{mapping}{aesthetic mapping}

\item{data}{data}

\item{layout}{one of 'rectangular', 'dendrogram', 'slanted', 'ellipse', 'roundrect',
'fan', 'circular', 'inward_circular', 'radial', 'equal_angle', 'daylight' or 'ape'}

\item{multiPhylo}{logical, whether input data contains multiple phylo class.}

\item{...}{additional parameter

some dot arguments:
\itemize{
\item \code{continuous} a character, which the aesthethic (\code{size} or \code{colour}) will be continuous. It
should be one of 'color', 'size', 'all' and 'NULL', default is NULL.
\item \code{nsplit} integer, the number of branch blocks divided when \code{continuous} is not NULL, default is 200.
default is TRUE, which is useful to 'aes(size=I(variable))'.
}}
}
\value{
tree layer
}
\description{
add tree layer
}
\section{Aesthetics}{

\code{geom_tree()} understands the following aesthethics:
\itemize{
\item \code{colour} logical, control the color of line, default is black.
\item \code{linetype} control the type of line, default is 1 (solid).
\item \code{size} numeric, control the width of line, default is 0.5.
}
}

\author{
Yu Guangchuang
}
