% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tree.R
\name{geom_tree}
\alias{geom_tree}
\title{geom_tree}
\usage{
geom_tree(
  mapping = NULL,
  data = NULL,
  layout = "rectangular",
  multiPhylo = FALSE,
  ...
)
}
\arguments{
\item{mapping}{aesthetic mapping}

\item{data}{data}

\item{layout}{one of 'rectangular', 'dendrogram', 'slanted', 'ellipse', 'roundrect',
'fan', 'circular', 'inward_circular', 'radial', 'equal_angle', 'daylight' or 'ape'}

\item{multiPhylo}{logical}

\item{...}{additional parameter}
}
\value{
tree layer
}
\description{
add tree layer
}
\author{
Yu Guangchuang
}
