% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{geom_tree}
\alias{geom_tree}
\title{geom_tree}
\usage{
geom_tree(layout = "rectangular", ...)
}
\arguments{
\item{layout}{one of 'rectangular', 'slanted', 'circular', 'radial' or 'unrooted'}

\item{...}{additional parameter}
}
\value{
tree layer
}
\description{
add tree layer
}
\examples{
require(ape)
tr <- rtree(10)
require(ggplot2)
ggplot(tr) + geom_tree()
}
\author{
Yu Guangchuang
}

