% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_treescale.R
\name{geom_treescale}
\alias{geom_treescale}
\title{geom_treescale}
\usage{
geom_treescale(x = NULL, y = NULL, width = NULL, offset = NULL,
  color = "black", linesize = 0.5, fontsize = 3.88)
}
\arguments{
\item{x}{x position}

\item{y}{y position}

\item{width}{width of scale}

\item{offset}{offset of text to line}

\item{color}{color}

\item{linesize}{size of line}

\item{fontsize}{size of text}
}
\value{
ggplot layers
}
\description{
add tree scale
}
\author{
Guangchuang Yu
}

