% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-zoom-clade.R
\name{geom_zoom_clade}
\alias{geom_zoom_clade}
\title{geom_zoom_clade}
\usage{
geom_zoom_clade(node, xexpand = NULL)
}
\arguments{
\item{node}{internal node number to zoom in its corresponding clade}

\item{xexpand}{numeric, extend x, meaning the ratio of range of the xlim of the original tree,
defaults to NULL.}
}
\value{
updated tree view
}
\description{
zoom selected clade of a tree
}
\details{
'geom_zoom_clade' zooms in on a selected clade of a tree,
while showing its on the full view of tree as a seperated panel for reference
}
\author{
Guangchuang Yu
}
