% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/beast.R, R/jplace.R
\docType{methods}
\name{get.fields}
\alias{get.fields}
\alias{get.fields,beast-method}
\alias{get.fields,jplace,ANY-method}
\title{get.fields method}
\usage{
get.fields(object, ...)

\S4method{get.fields}{beast}(object, ...)

get.fields(object, ...)
}
\arguments{
\item{object}{one of \code{jplace}, \code{beast} object}

\item{...}{additional parameter}
}
\value{
available annotation variables
}
\description{
get.fields method

get.fields method
}
\examples{
jp <- system.file("extdata", "sample.jplace", package="ggtree")
jp <- read.jplace(jp)
get.fields(jp)
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

