% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/RAxML.R, R/beast.R, R/codeml.R, R/codeml_mlc.R, R/hyphy.R, R/jplace.R, R/paml_rst.R, R/r8s.R
\docType{methods}
\name{get.fields}
\alias{get.fields}
\alias{get.fields,beast-method}
\alias{get.fields,codeml-method}
\alias{get.fields,codeml_mlc-method}
\alias{get.fields,hyphy-method}
\alias{get.fields,jplace,ANY-method}
\alias{get.fields,paml_rst-method}
\alias{get.fields,r8s-method}
\alias{get.fields,raxml-method}
\title{get.fields method}
\usage{
get.fields(object, ...)

\S4method{get.fields}{raxml}(object, ...)

\S4method{get.fields}{beast}(object, ...)

\S4method{get.fields}{codeml}(object, ...)

\S4method{get.fields}{codeml_mlc}(object)

\S4method{get.fields}{hyphy}(object, ...)

get.fields(object, ...)

\S4method{get.fields}{paml_rst}(object)

\S4method{get.fields}{r8s}(object, ...)
}
\arguments{
\item{object}{one of \code{jplace}, \code{beast}, \code{hyphy}, \code{codeml}, \code{codeml_mlc}, \code{paml_rst} object}

\item{...}{additional parameter}
}
\value{
available annotation variables
}
\description{
get.fields method

get.fields method
}
\examples{
jp <- system.file("extdata", "sample.jplace", package="ggtree")
jp <- read.jplace(jp)
get.fields(jp)
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

