% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree.R
\name{get.offspring.tip}
\alias{get.offspring.tip}
\title{get.offspring.tip}
\usage{
get.offspring.tip(tr, node)
}
\arguments{
\item{tr}{tree}

\item{node}{node}
}
\value{
tip label
}
\description{
extract offspring tips
}
\author{
ygc
}

