% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree.R
\name{get.path}
\alias{get.path}
\title{get.path}
\usage{
get.path(phylo, from, to)
}
\arguments{
\item{phylo}{phylo object}

\item{from}{start node}

\item{to}{end node}
}
\value{
node vectot
}
\description{
path from start node to end node
}
\author{
Guangchuang Yu
}

