% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/codeml.R, R/hyphy.R, R/paml_rst.R
\docType{methods}
\name{get.subs}
\alias{get.subs}
\alias{get.subs,codeml-method}
\alias{get.subs,hyphy-method}
\alias{get.subs,paml_rst-method}
\title{get.subs method}
\usage{
get.subs(object, type, ...)

\S4method{get.subs}{codeml}(object, type, ...)

\S4method{get.subs}{hyphy}(object, type, ...)

\S4method{get.subs}{paml_rst}(object, type, ...)
}
\arguments{
\item{object}{paml_rst object}

\item{type}{one of 'marginal_subs', 'marginal_AA_subs',
'joint_subs' or 'joint_AA_subs'.}

\item{...}{additional parameter}
}
\value{
data.frame
}
\description{
get.subs method
}
\examples{
nwk <- system.file("extdata/HYPHY", "labelledtree.tree", package="ggtree")
ancseq <- system.file("extdata/HYPHY", "ancseq.nex", package="ggtree")
tipfas <- system.file("extdata", "pa.fas", package="ggtree")
hy <- read.hyphy(nwk, ancseq, tipfas)
get.subs(hy, type="AA_subs")
}

