% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/beast.R
\docType{methods}
\name{get.tree}
\alias{get.tree}
\alias{get.tree,beast}
\title{get.tree method}
\usage{
get.tree(object, ...)

get.tree(object, ...)
}
\arguments{
\item{object}{one of \code{jplace}, \code{beast} object}

\item{...}{additional parameter}
}
\value{
phylo object
}
\description{
get.tree method

get.tree method
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

