% Generated by roxygen2 (4.1.0): do not edit by hand
\docType{methods}
\name{get.tree}
\alias{get.tree}
\alias{get.tree,jplace,ANY-method}
\title{get.tree method}
\usage{
get.tree(object, ...)

get.tree(object, ...)
}
\arguments{
\item{object}{jplace object}

\item{...}{additional parameter}
}
\value{
tree text
}
\description{
get.tree method

get.tree method
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

