% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/beast.R, R/codeml_mlc.R, R/paml_rst.R
\docType{methods}
\name{get.tree}
\alias{get.tree}
\alias{get.tree,beast}
\alias{get.tree,codeml_mlc-method}
\alias{get.tree,paml_rst-method}
\title{get.tree method}
\usage{
get.tree(object, ...)

get.tree(object, ...)

\S4method{get.tree}{codeml_mlc}(object, ...)

\S4method{get.tree}{paml_rst}(object)
}
\arguments{
\item{object}{one of \code{jplace}, \code{beast} object}

\item{...}{additional parameter}
}
\value{
phylo object
}
\description{
get.tree method

get.tree method
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

