% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/beast.R, R/jplace.R
\docType{methods}
\name{get.tree}
\alias{get.tree}
\alias{get.tree,beast-method}
\alias{get.tree,jplace,ANY-method}
\title{get.tree method}
\usage{
get.tree(object, ...)

\S4method{get.tree}{beast}(object, ...)

get.tree(object, ...)
}
\arguments{
\item{object}{jplace object}

\item{...}{additional parameter}
}
\value{
tree text
}
\description{
get.tree method

get.tree method
}
\examples{
jp <- system.file("extdata", "sample.jplace", package="ggtree")
jp <- read.jplace(jp)
get.tree(jp)
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

