% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/jplace.R
\docType{methods}
\name{get.treeinfo}
\alias{get.treeinfo}
\alias{get.treeinfo,jplace,ANY-method}
\title{get.treeinfo method}
\usage{
get.treeinfo(object, layout = "phylogram", ladderize = TRUE,
  right = FALSE, ...)

get.treeinfo(object, layout, ladderize, right, ...)
}
\arguments{
\item{object}{jplace object}

\item{layout}{layout}

\item{ladderize}{ladderize, logical}

\item{right}{logical, parameter for ladderize}

\item{...}{additional parameter}
}
\value{
data.frame
}
\description{
get.treeinfo method

get.treeinfo method
}
\examples{
jp <- system.file("extdata", "sample.jplace", package="ggtree")
jp <- read.jplace(jp)
get.treeinfo(jp)
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

