% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/jplace.R
\docType{methods}
\name{get.treetext}
\alias{get.treetext}
\alias{get.treetext,jplace,ANY-method}
\title{get.treetext method}
\usage{
get.treetext(object, ...)

get.treetext(object, ...)
}
\arguments{
\item{object}{one of \code{phylo}, \code{jplace}, \code{beast}, \code{hyphy}, \code{codeml}, \code{codeml_mlc}, \code{paml_rst} object}

\item{...}{additional parameter}
}
\value{
phylo object
}
\description{
get.treetext method
}
\examples{
jp <- system.file("extdata", "sample.jplace", package="ggtree")
jp <- read.jplace(jp)
get.treetext(jp)
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

