% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade-functions.R
\name{get_taxa_name}
\alias{get_taxa_name}
\title{get_taxa_name}
\usage{
get_taxa_name(tree_view = NULL, node = NULL)
}
\arguments{
\item{tree_view}{tree view (i.e. the ggtree object). If tree_view is NULL, the last ggplot will be used.}

\item{node}{internal node number}
}
\value{
taxa name vector
}
\description{
get taxa name of a selected node (or tree if node=NULL) sorted by their position in plotting
}
\author{
Guangchuang Yu
}
