% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{get_taxa_name}
\alias{get_taxa_name}
\title{get_taxa_name}
\usage{
get_taxa_name(tree_view, node)
}
\arguments{
\item{tree_view}{tree view}

\item{node}{node}
}
\value{
taxa name vector
}
\description{
get taxa name of a selected node
}
\author{
Guangchuang Yu
}

