% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytree.R
\name{GetSubtree.df}
\alias{GetSubtree.df}
\title{GetSubtree.df}
\usage{
GetSubtree.df(df, node)
}
\arguments{
\item{df}{tree data.frame}

\item{node}{id of starting node.}
}
\value{
list of all child node id's from starting node.
}
\description{
Get all children of node from df tree using breath-first.
}
