% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdensitree.R
\name{ggdensitree}
\alias{ggdensitree}
\title{ggdensitree}
\usage{
ggdensitree(data = NULL, mapping = NULL, layout = "slanted",
  tip.order = "mds", dangle = TRUE, jitter = 0, ...)
}
\arguments{
\item{data}{a list of phylo objects or any object with an as.phylo and fortify method}

\item{mapping}{aesthetic mapping}

\item{layout}{one of 'slanted', 'rectangluar', 'fan', 'circular' or 'radial' (default: 'slanted')}

\item{tip.order}{the order of the tips by a character vector of taxa names; or an integer, N, to order the tips by the order of the tips in the Nth tree; or 'mds' to order the tips based on MDS of the distance between the tips (default: 'mds')}

\item{dangle}{TRUE to align trees by their tips and FALSE to align treesby their root (default: TRUE)}

\item{jitter}{deviation to jitter tips}

\item{...}{additional parameters passed to fortify, ggtree and geom_tree}
}
\value{
tree layer
}
\description{
drawing phylogenetic trees from list of phylo objects
}
\examples{
require(ape)
require(dplyr)
library(ape)
library(tidyverse)

trees <- list(read.tree(text="((a,b),c);"), read.tree(text="((a,c),b);"));
ggdensitree(trees) + geom_tiplab()

trees.fort <- list(trees[[1]] \%>\% fortify \%>\% mutate(tree="a"), trees[[2]] \%>\% fortify \%>\% mutate(tree="b"));
ggdensitree(trees.fort, aes(colour=tree)) + geom_tiplab(colour='black')
}
\author{
Yu Guangchuang, Bradley R. Jones
}
