% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggtree}
\alias{ggtree}
\title{ggtree}
\usage{
ggtree(tr, dist.legend = FALSE, ...)
}
\arguments{
\item{tr}{phylo object}

\item{dist.legend}{add distance legend, logical}

\item{...}{additional parameter}
}
\value{
tree
}
\description{
drawing phylogenetic tree from phylo object
}
\author{
Yu Guangchuang
}

