% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggtree-package.R, R/ggtree.R
\docType{package}
\name{ggtree}
\alias{ggtree}
\alias{ggtree-package}
\alias{package-ggtree}
\title{visualizing phylogenetic tree and heterogenous associated data based on grammar of graphics
\code{ggtree} provides functions for visualizing phylogenetic tree and its associated data in R.}
\usage{
ggtree(tr, showDistance = FALSE, layout = "phylogram", yscale = "none",
  ladderize = TRUE, right = FALSE, branch.length = "branch.length",
  ndigits = NULL, ...)
}
\arguments{
\item{tr}{phylo object}

\item{showDistance}{add distance legend, logical}

\item{layout}{one of phylogram, dendrogram, cladogram, fan, radial and unrooted}

\item{yscale}{y scale}

\item{ladderize}{logical}

\item{right}{logical}

\item{branch.length}{variable for scaling branch}

\item{ndigits}{number of digits to round numerical annotation variable}

\item{...}{additional parameter}
}
\value{
tree
}
\description{
visualizing phylogenetic tree and heterogenous associated data based on grammar of graphics
\code{ggtree} provides functions for visualizing phylogenetic tree and its associated data in R.

drawing phylogenetic tree from phylo object
}
\examples{
require(ape)
tr <- rtree(10)
ggtree(tr)
}
\author{
Yu Guangchuang
}

