% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtree-package.R, R/ggtree.R
\docType{package}
\name{ggtree}
\alias{ggtree}
\alias{package-ggtree}
\alias{ggtree-package}
\title{visualizing phylogenetic tree and heterogenous associated data based on grammar of graphics
\code{ggtree} provides functions for visualizing phylogenetic tree and its associated data in R.}
\usage{
ggtree(tr, mapping = NULL, layout = "rectangular", open.angle = 0,
  mrsd = NULL, as.Date = FALSE, yscale = "none",
  yscale_mapping = NULL, ladderize = TRUE, right = FALSE,
  branch.length = "branch.length", ...)
}
\arguments{
\item{tr}{phylo object}

\item{mapping}{aes mapping}

\item{layout}{one of 'rectangular', 'slanted', 'fan', 'circular', 'radial', 'equal_angle' or 'daylight'}

\item{open.angle}{open angle, only for 'fan' layout}

\item{mrsd}{most recent sampling date}

\item{as.Date}{logical whether using Date class in time tree}

\item{yscale}{y scale}

\item{yscale_mapping}{yscale mapping for category variable}

\item{ladderize}{logical (default \code{TRUE}). Should the tree be re-organized to have a 'ladder'
aspect?}

\item{right}{logical. If \code{ladderize = TRUE}, should the ladder have the smallest clade on the
right-hand side? See \code{\link[ape]{ladderize}} for more information.}

\item{branch.length}{variable for scaling branch, if 'none' draw cladogram}

\item{...}{additional parameter}
}
\value{
tree
}
\description{
If you use ggtree in published research, please cite:
Guangchuang Yu, David Smith, Huachen Zhu, Yi Guan, Tommy Tsan-Yuk Lam.
ggtree: an R package for visualization and annotation of phylogenetic trees with their covariates and other associated data.
Methods in Ecology and Evolution 2017, 8(1):28-36, doi:10.1111/2041-210X.12628

drawing phylogenetic tree from phylo object
}
\examples{
require(ape)
tr <- rtree(10)
ggtree(tr)
}
\seealso{
\code{\link[ape]{ladderize}}
}
\author{
Yu Guangchuang
}
