% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggtree-package.R, R/ggtree.R
\docType{package}
\name{ggtree}
\alias{ggtree}
\alias{ggtree-package}
\alias{package-ggtree}
\title{visualizing phylogenetic tree and heterogenous associated data based on grammar of graphics
\code{ggtree} provides functions for visualizing phylogenetic tree and its associated data in R.}
\usage{
ggtree(tr, showDistance = FALSE, ...)
}
\arguments{
\item{tr}{phylo object}

\item{showDistance}{add distance legend, logical}

\item{...}{additional parameter}
}
\value{
tree
}
\description{
visualizing phylogenetic tree and heterogenous associated data based on grammar of graphics
\code{ggtree} provides functions for visualizing phylogenetic tree and its associated data in R.

drawing phylogenetic tree from phylo object
}
\examples{
require(ape)
tr <- rtree(10)
ggtree(tr)
}
\author{
Yu Guangchuang
}

