% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtree-package.R, R/ggtree.R
\docType{package}
\name{ggtree}
\alias{ggtree}
\alias{package-ggtree}
\alias{ggtree-package}
\alias{ggtree}
\title{visualizing phylogenetic tree and heterogenous associated data based on grammar of graphics
\code{ggtree} provides functions for visualizing phylogenetic tree and its associated data in R.}
\usage{
ggtree(tr, mapping = NULL, layout = "rectangular", open.angle = 0,
  mrsd = NULL, as.Date = FALSE, yscale = "none", yscale_mapping = NULL,
  ladderize = TRUE, right = FALSE, branch.length = "branch.length",
  ndigits = NULL, ...)
}
\arguments{
\item{tr}{phylo object}

\item{mapping}{aes mapping}

\item{layout}{one of 'rectangular', 'slanted', 'fan', 'circular', 'radial', 'equal_angle' or 'daylight'}

\item{open.angle}{open angle, only for 'fan' layout}

\item{mrsd}{most recent sampling date}

\item{as.Date}{logical whether using Date class in time tree}

\item{yscale}{y scale}

\item{yscale_mapping}{yscale mapping for category variable}

\item{ladderize}{logical}

\item{right}{logical}

\item{branch.length}{variable for scaling branch, if 'none' draw cladogram}

\item{ndigits}{number of digits to round numerical annotation variable}

\item{...}{additional parameter}
}
\value{
tree
}
\description{
visualizing phylogenetic tree and heterogenous associated data based on grammar of graphics
\code{ggtree} provides functions for visualizing phylogenetic tree and its associated data in R.

drawing phylogenetic tree from phylo object
}
\examples{
require(ape)
tr <- rtree(10)
ggtree(tr)
}
\author{
Yu Guangchuang
}
