% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-print.r
\name{ggtree_set_interactive}
\alias{ggtree_set_interactive}
\title{Set ggtree interactive mode}
\usage{
ggtree_set_interactive(check = TRUE)
}
\arguments{
\item{check}{whether to check interactive attributes}
}
\description{
set ggtree interactive mode
}
