% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gheatmap.R
\name{gheatmap}
\alias{gheatmap}
\title{gheatmap}
\usage{
gheatmap(p, data, offset = 0, width = 1, low = "green", high = "red",
  color = "white", colnames = TRUE, colnames_position = "bottom",
  colnames_angle = 0, colnames_level = NULL, colnames_offset_x = 0,
  colnames_offset_y = 0, font.size = 4, hjust = 0.5)
}
\arguments{
\item{p}{tree view}

\item{data}{matrix or data.frame}

\item{offset}{offset of heatmap to tree}

\item{width}{total width of heatmap, compare to width of tree}

\item{low}{color of lowest value}

\item{high}{color of highest value}

\item{color}{color of heatmap cell border}

\item{colnames}{logical, add matrix colnames or not}

\item{colnames_position}{one of 'bottom' or 'top'}

\item{colnames_angle}{angle of column names}

\item{colnames_level}{levels of colnames}

\item{colnames_offset_x}{x offset for column names}

\item{colnames_offset_y}{y offset for column names}

\item{font.size}{font size of matrix colnames}

\item{hjust}{hjust for column names (0: align left, 0.5: align center, 1: align righ)}
}
\value{
tree view
}
\description{
append a heatmap of a matrix to right side of phylogenetic tree
}
\author{
Guangchuang Yu
}
