% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gplot.R
\name{gplot}
\alias{gplot}
\title{view tree and associated matrix}
\usage{
gplot(p, data, low = "green", high = "red", widths = c(0.5, 0.5),
  font.size = 14)
}
\arguments{
\item{p}{tree view}

\item{data}{matrix}

\item{low}{low color}

\item{high}{high color}

\item{widths}{widths of sub plot}

\item{font.size}{font size}
}
\description{
view tree and associated matrix
}

