% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-groupClade.R
\docType{methods}
\name{groupClade}
\alias{groupClade}
\alias{groupClade,ggtree-method}
\title{groupClade method}
\usage{
\S4method{groupClade}{ggtree}(object, node, group_name)
}
\arguments{
\item{object}{ggtree object}

\item{node}{internal node number}

\item{group_name}{name of the group}
}
\description{
groupClade method for ggtree object
}

