% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/RAxML.R, R/hyphy.R, R/method-groupOTU.R
\docType{methods}
\name{groupOTU}
\alias{groupOTU}
\alias{groupOTU,apeBootstrap-method}
\alias{groupOTU,beast-method}
\alias{groupOTU,codeml-method}
\alias{groupOTU,codeml_mlc-method}
\alias{groupOTU,gg-method}
\alias{groupOTU,ggplot-method}
\alias{groupOTU,hyphy-method}
\alias{groupOTU,jplace-method}
\alias{groupOTU,nhx-method}
\alias{groupOTU,paml_rst-method}
\alias{groupOTU,phangorn-method}
\alias{groupOTU,phylip-method}
\alias{groupOTU,phylo-method}
\alias{groupOTU,r8s-method}
\alias{groupOTU,raxml-method}
\title{groupOTU method}
\usage{
groupOTU(object, focus, group_name = "group", ...)

\S4method{groupOTU}{raxml}(object, focus, group_name = "group")

\S4method{groupOTU}{hyphy}(object, focus, group_name = "group")

\S4method{groupOTU}{apeBootstrap}(object, focus, group_name = "group")

\S4method{groupOTU}{beast}(object, focus, group_name = "group")

\S4method{groupOTU}{codeml}(object, focus, group_name = "group")

\S4method{groupOTU}{codeml_mlc}(object, focus, group_name = "group")

\S4method{groupOTU}{gg}(object, focus, group_name)

\S4method{groupOTU}{ggplot}(object, focus, group_name = "group")

\S4method{groupOTU}{jplace}(object, focus, group_name = "group")

\S4method{groupOTU}{nhx}(object, focus, group_name = "group")

\S4method{groupOTU}{phangorn}(object, focus, group_name = "group")

\S4method{groupOTU}{phylip}(object, focus, group_name = "group")

\S4method{groupOTU}{paml_rst}(object, focus, group_name = "group")

\S4method{groupOTU}{phylo}(object, focus, group_name = "group")

\S4method{groupOTU}{r8s}(object, focus, group_name = "group", tree = "TREE")
}
\arguments{
\item{object}{supported objects, including phylo, paml_rst,
codeml_mlc, codeml, jplace, beast, hyphy}

\item{focus}{a vector of tip (label or number) or a list of tips.}

\item{group_name}{name of the group, 'group' by default}

\item{...}{additional parameter}

\item{tree}{which tree selected}
}
\value{
group index
}
\description{
group tree based on selected OTU, will traceback to MRCA
}

