% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phylo.R
\name{groupOTU.phylo}
\alias{groupOTU.phylo}
\title{groupOTU.phylo}
\usage{
groupOTU.phylo(phy, focus, group_name = "group")
}
\arguments{
\item{phy}{tree object}

\item{focus}{tip list}

\item{group_name}{name of the group}
}
\value{
phylo object
}
\description{
group OTU
}
\author{
ygc
}

