% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-groupOTU.R
\name{groupOTU.phylo}
\alias{groupOTU.phylo}
\title{groupOTU.phylo}
\usage{
groupOTU.phylo(phy, focus, group_name = "group", ...)
}
\arguments{
\item{phy}{tree object}

\item{focus}{tip list}

\item{group_name}{name of the group}

\item{...}{additional parameters}
}
\value{
phylo object
}
\description{
group OTU
}
\author{
ygc
}

