% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/defunct.R
\docType{methods}
\name{gzoom}
\alias{gzoom}
\alias{gzoom,ggtree-method}
\alias{gzoom,treedata-method}
\alias{gzoom,phylo-method}
\title{gzoom method}
\usage{
gzoom(object, focus, subtree = FALSE, widths = c(0.3, 0.7), ...)

\S4method{gzoom}{ggtree}(object, focus, widths = c(0.3, 0.7), xmax_adjust = 0)

\S4method{gzoom}{treedata}(object, focus, subtree = FALSE, widths = c(0.3, 0.7))

\S4method{gzoom}{phylo}(object, focus, subtree = FALSE, widths = c(0.3, 0.7))
}
\arguments{
\item{object}{supported tree objects}

\item{focus}{selected tips}

\item{subtree}{logical}

\item{widths}{widths}

\item{...}{additional parameter}

\item{xmax_adjust}{adjust xmax (\code{xlim[2]})}
}
\value{
figure
}
\description{
gzoom method

gzoom method

zoom selected subtree
}
