% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{hilight}
\alias{hilight}
\title{hilight}
\usage{
hilight(tree_view, node, fill = "steelblue", alpha = 0.5, ...)
}
\arguments{
\item{tree_view}{tree view}

\item{node}{clade node}

\item{fill}{fill color}

\item{alpha}{alpha}

\item{...}{additional parameter}
}
\value{
tree view
}
\description{
hilight clade with rectangle
}
\author{
Guangchuang Yu
}

