% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-identify.R
\name{identify.gg}
\alias{identify.gg}
\title{identify}
\usage{
\method{identify}{gg}(x, ...)
}
\arguments{
\item{x}{tree view}

\item{...}{additional parameters}
}
\value{
node id
}
\description{
identify node by interactive click
}
\author{
Guangchuang Yu
}

