% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset.R
\name{inset}
\alias{inset}
\title{inset}
\usage{
inset(tree_view, insets, width = 0.1, height = 0.1, hjust = 0,
  vjust = 0, x = "node")
}
\arguments{
\item{tree_view}{tree view}

\item{insets}{a list of ggplot objects, named by node number}

\item{width}{width of inset}

\item{height}{height of inset}

\item{hjust}{horizontal adjustment}

\item{vjust}{vertical adjustment}

\item{x}{x position, one of 'node' and 'branch'}
}
\value{
tree view with insets
}
\description{
add insets in a tree
}
\author{
Guangchuang Yu
}

