% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{jplace-class}
\alias{get.fields,jplace-method}
\alias{get.placements,jplace-method}
\alias{get.treeinfo,jplace-method}
\alias{get.treetext,jplace-method}
\alias{jplace-class}
\alias{show,jplace-method}
\title{Class "jplace"
This class stores information of jplace file.}
\description{
Class "jplace"
This class stores information of jplace file.
}
\section{Slots}{

\describe{
\item{\code{fields}}{colnames of first variable of placements}

\item{\code{treetext}}{tree text}

\item{\code{phylo}}{tree phylo object}

\item{\code{placements}}{placement information}

\item{\code{version}}{version}

\item{\code{metadata}}{metadata}

\item{\code{file}}{jplace file}

\item{\code{extraInfo}}{extra information}
}}
\author{
Guangchuang Yu \url{http://guangchuangyu.github.io}
}
\seealso{
\code{\link{show}} \code{\link{get.tree}}
         \code{\link{ggtree}}
}
\keyword{classes}

