% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gplot.R
\name{msaplot}
\alias{msaplot}
\title{msaplot}
\usage{
msaplot(p, fasta, offset = 0, width = NULL, color = NULL, window = NULL)
}
\arguments{
\item{p}{tree view}

\item{fasta}{fasta file, multiple sequence alignment}

\item{offset}{offset of MSA to tree}

\item{width}{width of each character}

\item{color}{color}

\item{window}{specific a slice to display}
}
\value{
tree view
}
\description{
multiple sequence alignment with phylogenetic tree
}
\author{
Guangchuang Yu
}

