% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{multiplot}
\usage{
multiplot(..., plotlist = NULL, ncol, widths = rep_len(1, ncol),
  labels = NULL, label_size = 5)
}
\arguments{
\item{...}{plots}

\item{plotlist}{plot list}

\item{ncol}{number of column}

\item{widths}{widths of plots}

\item{labels}{labels for labeling the plots}

\item{label_size}{font size of label}
}
\value{
plot
}
\description{
plot multiple ggplot objects in one page
}
\author{
Guangchuang Yu
}

