% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset.R
\name{nodebar}
\alias{nodebar}
\title{nodebar}
\usage{
nodebar(data, cols, color, alpha = 1, position = "stack")
}
\arguments{
\item{data}{a data.frame of stats with an additional column of node number}

\item{cols}{column of stats}

\item{color}{color of bar}

\item{alpha}{alpha}

\item{position}{position of bar, one of 'stack' and 'dodge'}
}
\value{
list of ggplot objects
}
\description{
generate a list of bar charts for results of ancestral state reconstruction
}
\author{
Guangchuang Yu
}
