% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{open_tree}
\alias{open_tree}
\title{open_tree}
\usage{
open_tree(treeview, angle)
}
\arguments{
\item{treeview}{tree view}

\item{angle}{angle}
}
\value{
updated tree view
}
\description{
open tree with specific angle
}
\author{
Guangchuang Yu
}
