% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/beast.R, R/codeml.R, R/codeml_mlc.R, R/hyphy.R, R/paml_rst.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,beast,ANY-method}
\alias{plot,codeml,ANY-method}
\alias{plot,codeml_mlc,ANY-method}
\alias{plot,hyphy,ANY-method}
\alias{plot,paml_rst,ANY-method}
\title{plot method}
\usage{
\S4method{plot}{beast,ANY}(x, layout = "phylogram",
  branch.length = "branch.length", show.tip.label = TRUE,
  tip.label.size = 4, tip.label.hjust = -0.1, position = "branch",
  annotation = "rate", ndigits = 2, annotation.size = 3,
  annotation.color = "black", ...)

\S4method{plot}{codeml,ANY}(x, layout = "phylogram",
  branch.length = "mlc.branch.length", show.tip.label = TRUE,
  tip.label.size = 4, tip.label.hjust = -0.1, position = "branch",
  annotation = "dN_vs_dS", annotation.size = 3,
  annotation.color = "black", ndigits = 2, ...)

\S4method{plot}{codeml_mlc,ANY}(x, layout = "phylogram",
  branch.length = "branch.length", show.tip.label = TRUE,
  tip.label.size = 4, tip.label.hjust = -0.1, position = "branch",
  annotation = "dN_vs_dS", annotation.size = 3,
  annotation.color = "black", ndigits = 2, ...)

\S4method{plot}{hyphy,ANY}(x, layout = "phylogram", show.tip.label = TRUE,
  tip.label.size = 4, tip.label.hjust = -0.1, position = "branch",
  annotation = "subs", annotation.color = "black", annotation.size = 3,
  ...)

\S4method{plot}{paml_rst,ANY}(x, layout = "phylogram",
  show.tip.label = TRUE, tip.label.size = 4, tip.label.hjust = -0.1,
  position = "branch", annotation = "marginal_subs",
  annotation.color = "black", annotation.size = 3, ...)
}
\arguments{
\item{x}{object}

\item{layout}{layout}

\item{branch.length}{branch length}

\item{show.tip.label}{logical}

\item{tip.label.size}{size of tip label}

\item{tip.label.hjust}{hjust of tip.label}

\item{position}{one of "branch" and "node"}

\item{annotation}{one of get.fields(x)}

\item{ndigits}{round digits}

\item{annotation.size}{size of annotation}

\item{annotation.color}{color of annotation}

\item{...}{Additional argument list}
}
\value{
plot
}
\description{
plot method generics
}
\examples{
file <- system.file("extdata/BEAST", "beast_mcc.tree", package="ggtree")
beast <- read.beast(file)
plot(beast, annotation="length_0.95_HPD", branch.length="none") + theme_tree()
}

