% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/beast.R, R/codeml.R, R/codeml_mlc.R, R/paml_rst.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,beast,ANY-method}
\alias{plot,codeml,ANY-method}
\alias{plot,codeml_mlc,ANY-method}
\alias{plot,paml_rst,ANY-method}
\title{plot method}
\usage{
\S4method{plot}{beast,ANY}(x, layout = "phylogram",
  branch.length = "branch.length", show.tip.label = TRUE,
  position = "branch", annotation = "rate", ndigits = 2, ...)

\S4method{plot}{codeml,ANY}(x, layout = "phylogram",
  branch.length = "mlc.branch.length", show.tip.label = TRUE,
  position = "branch", annotation = "dN/dS", ndigits = 2, ...)

\S4method{plot}{codeml_mlc,ANY}(x, layout = "phylogram",
  branch.length = "branch.length", show.tip.label = TRUE,
  position = "branch", annotation = "dN/dS", ndigits = 2, ...)

\S4method{plot}{paml_rst,ANY}(x, layout = "phylogram",
  show.tip.label = TRUE, position = "branch",
  annotation = "marginal_subs", ...)
}
\arguments{
\item{x}{object}

\item{layout}{layout}

\item{branch.length}{branch length}

\item{show.tip.label}{logical}

\item{position}{one of "branch" and "node"}

\item{annotation}{one of get.fields(x)}

\item{ndigits}{round digits}

\item{...}{Additional argument list}
}
\value{
plot
}
\description{
plot method generics
}

