% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/codeml_mlc.R
\name{read.codeml_mlc}
\alias{read.codeml_mlc}
\title{read.codeml_mlc}
\usage{
read.codeml_mlc(mlcfile)
}
\arguments{
\item{mlcfile}{mlc file}
}
\value{
A \code{codeml_mlc} object
}
\description{
read mlc file of codeml output
}
\examples{
mlcfile <- system.file("extdata/PAML_Codeml", "mlc", package="ggtree")
read.codeml_mlc(mlcfile)
}
\author{
ygc
}

