% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHX.R
\name{read.nhx}
\alias{read.nhx}
\title{read.nhx}
\usage{
read.nhx(file)
}
\arguments{
\item{file}{nhx file}
}
\value{
nhx object
}
\description{
read nhx tree file
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

