% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paml_rst.R
\name{read.paml_rst}
\alias{read.paml_rst}
\title{read.paml_rst}
\usage{
read.paml_rst(rstfile)
}
\arguments{
\item{rstfile}{rst file}
}
\value{
A \code{paml_rst} object
}
\description{
read rst file from paml output
}
\examples{
rstfile <- system.file("extdata/PAML_Baseml", "rst", package="ggtree")
read.paml_rst(rstfile)
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

