% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{rtree}
\alias{reexports}
\alias{read.tree}
\alias{reexports}
\alias{groupOTU}
\alias{reexports}
\alias{groupClade}
\alias{reexports}
\alias{xlim}
\alias{reexports}
\alias{theme}
\alias{reexports}
\alias{ggsave}
\alias{reexports}
\alias{aes}
\alias{reexports}
\alias{geom_text}
\alias{reexports}
\alias{geom_label}
\alias{reexports}
\alias{geom_point}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ape}{\code{\link[ape]{rtree}}, \code{\link[ape]{read.tree}}}

  \item{ggplot2}{\code{\link[ggplot2]{xlim}}, \code{\link[ggplot2]{theme}}, \code{\link[ggplot2]{ggsave}}, \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{geom_text}}, \code{\link[ggplot2]{geom_label}}, \code{\link[ggplot2]{geom_point}}}

  \item{tidytree}{\code{\link[tidytree]{groupOTU}}, \code{\link[tidytree]{groupClade}}}
}}

