% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{rtree}
\alias{read.tree}
\alias{read.nexus}
\alias{groupOTU}
\alias{groupClade}
\alias{collapse}
\alias{fortify}
\alias{ggplot}
\alias{xlim}
\alias{theme}
\alias{ggsave}
\alias{aes}
\alias{geom_text}
\alias{geom_label}
\alias{geom_point}
\alias{read.astral}
\alias{read.beast}
\alias{read.codeml}
\alias{read.codeml_mlc}
\alias{read.hyphy}
\alias{read.iqtree}
\alias{read.jplace}
\alias{read.jtree}
\alias{read.mrbayes}
\alias{read.newick}
\alias{read.nhx}
\alias{read.paml_rst}
\alias{read.phylip}
\alias{read.phyloT}
\alias{read.r8s}
\alias{read.raxml}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ape}{\code{\link[ape]{rtree}}, \code{\link[ape]{read.tree}}, \code{\link[ape]{read.nexus}}}

  \item{dplyr}{\code{\link[dplyr]{collapse}}}

  \item{ggplot2}{\code{\link[ggplot2]{fortify}}, \code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{xlim}}, \code{\link[ggplot2]{theme}}, \code{\link[ggplot2]{ggsave}}, \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{fortify}}, \code{\link[ggplot2]{geom_text}}, \code{\link[ggplot2]{geom_label}}, \code{\link[ggplot2]{geom_point}}}

  \item{tidytree}{\code{\link[tidytree]{groupOTU}}, \code{\link[tidytree]{groupClade}}}

  \item{treeio}{\code{\link[treeio]{read.astral}}, \code{\link[treeio]{read.beast}}, \code{\link[treeio]{read.codeml}}, \code{\link[treeio]{read.codeml_mlc}}, \code{\link[treeio]{read.hyphy}}, \code{\link[treeio]{read.iqtree}}, \code{\link[treeio]{read.jplace}}, \code{\link[treeio]{read.jtree}}, \code{\link[treeio]{read.mrbayes}}, \code{\link[treeio]{read.newick}}, \code{\link[treeio]{read.nhx}}, \code{\link[treeio]{read.paml_rst}}, \code{\link[treeio]{read.phylip}}, \code{\link[treeio]{read.phyloT}}, \code{\link[treeio]{read.r8s}}, \code{\link[treeio]{read.raxml}}}
}}

