% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape.R
\docType{import}
\name{reexports}
\alias{read.tree}
\alias{reexports}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ape}{\code{\link[ape]{read.tree}}}
}}
\keyword{internal}

