% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade-functions.R
\name{rotate}
\alias{rotate}
\title{rotate}
\usage{
rotate(tree_view = NULL, node)
}
\arguments{
\item{tree_view}{tree view}

\item{node}{selected node}
}
\value{
ggplot2 object
}
\description{
rotate 180 degree of a selected branch
}
\author{
Guangchuang Yu
}

