% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{rotate_tree}
\alias{rotate_tree}
\title{rotate_tree}
\usage{
rotate_tree(treeview, angle)
}
\arguments{
\item{treeview}{tree view}

\item{angle}{angle}
}
\value{
updated tree view
}
\description{
rotate circular tree
}
\author{
Guangchuang Yu
}

