% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/method-scale-color.R
\docType{methods}
\name{scale_color}
\alias{scale_color}
\alias{scale_color,beast-method}
\alias{scale_color,paml_rst-method}
\alias{scale_color,phylo-method}
\alias{scale_color,treedata-method}
\title{scale_color method}
\usage{
scale_color(object, by, ...)

\S4method{scale_color}{beast}(object, by, ...)

\S4method{scale_color}{treedata}(object, by, ...)

\S4method{scale_color}{paml_rst}(object, by, ...)

\S4method{scale_color}{phylo}(object, by, ...)
}
\arguments{
\item{object}{supported objects, including phylo, paml_rst,
codeml_mlc, codeml, jplace, beast, hyphy}

\item{by}{one of numerical attributes}

\item{...}{additional parameter}
}
\value{
color vector
}
\description{
scale color by a numerical tree attribute
}

