% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplot.R
\name{scale_x_ggtree}
\alias{scale_x_ggtree}
\title{scale_x_ggtree}
\usage{
scale_x_ggtree(tree_view, breaks = NULL, labels = NULL)
}
\arguments{
\item{tree_view}{tree view}

\item{breaks}{breaks for tree}

\item{labels}{lables for corresponding breaks}
}
\value{
tree view
}
\description{
scale x for tree with heatmap
}
\author{
Guangchuang Yu
}

