% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade-functions.R
\name{scaleClade}
\alias{scaleClade}
\title{scaleClade}
\usage{
scaleClade(tree_view = NULL, node, scale = 1, vertical_only = TRUE)
}
\arguments{
\item{tree_view}{tree view}

\item{node}{clade node}

\item{scale}{scale}

\item{vertical_only}{logical. If TRUE, only vertical will be scaled.
If FALSE, the clade will be scaled vertical and horizontally.
TRUE by default.}
}
\value{
tree view
}
\description{
scale clade
}
\author{
Guangchuang Yu
}
