% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/jplace.R
\docType{methods}
\name{"set.treeinfo<-"}
\alias{"set.treeinfo<-"}
\alias{set.treeinfo<-}
\alias{set.treeinfo<-,jplace,ANY-method}
\title{"set.treeinfo<-" method}
\usage{
set.treeinfo(x) <- value

set.treeinfo(x) <- value
}
\arguments{
\item{x}{one of \code{jplace}, \code{beast} object}

\item{value}{tree info}
}
\value{
updated jplace object
}
\description{
"set.treeinfo<-" method

set.treeinfo method for \code{jplace} object
}
\examples{
jp <- system.file("extdata", "sample.jplace", package="ggtree")
jp <- read.jplace(jp)
set.treeinfo(jp) <- get.treeinfo(jp)
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

