% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/jplace.R
\docType{methods}
\name{"set.treeinfo<-"}
\alias{"set.treeinfo<-"}
\alias{set.treeinfo<-}
\alias{set.treeinfo<-,jplace,ANY-method}
\title{"set.treeinfo<-" method}
\usage{
set.treeinfo(x) <- value

set.treeinfo(x) <- value
}
\arguments{
\item{x}{jplace object}

\item{value}{tree info}
}
\description{
"set.treeinfo<-" method

set.treeinfo method for \code{jplace} object
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

