% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{set_hilight_legend}
\alias{set_hilight_legend}
\title{set_hilight_legend}
\usage{
set_hilight_legend(p, color, label, alpha = 1)
}
\arguments{
\item{p}{ggtree object}

\item{color}{color vector}

\item{label}{label vector}

\item{alpha}{transparency of color}
}
\value{
updated ggtree object
}
\description{
set legend for multiple geom_hilight layers
}
\author{
Guangchuang Yu
}
