% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jplace.R
\docType{methods}
\name{show}
\alias{show}
\title{show method}
\usage{
show(object)
}
\arguments{
\item{object}{A \code{jplace} instance.}
}
\value{
print info
}
\description{
show method for \code{jplace} instance
}
\examples{
jp <- system.file("extdata", "sample.jplace", package="ggtree")
jp <- read.jplace(jp)
show(jp)
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

