% Generated by roxygen2 (4.1.0): do not edit by hand
\docType{methods}
\name{show}
\alias{show}
\title{show method}
\usage{
show(object)
}
\arguments{
\item{object}{A \code{jplace} instance.}
}
\value{
print info
}
\description{
show method for \code{jplace} instance
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

