% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAxML.R, R/ape.R, R/hyphy.R, R/method-show.R, R/phangorn.R
\docType{methods}
\name{show,raxml-method}
\alias{show}
\alias{show,apeBootstrap-method}
\alias{show,beast-method}
\alias{show,codeml-method}
\alias{show,codeml_mlc-method}
\alias{show,hyphy-method}
\alias{show,nhx-method}
\alias{show,paml_rst-method}
\alias{show,phangorn-method}
\alias{show,phylip-method}
\alias{show,r8s-method}
\alias{show,raxml-method}
\title{show method}
\usage{
\S4method{show}{raxml}(object)

\S4method{show}{apeBootstrap}(object)

\S4method{show}{hyphy}(object)

\S4method{show}{beast}(object)

\S4method{show}{codeml}(object)

\S4method{show}{codeml_mlc}(object)

show(object)

\S4method{show}{nhx}(object)

\S4method{show}{phylip}(object)

\S4method{show}{paml_rst}(object)

\S4method{show}{r8s}(object)

\S4method{show}{phangorn}(object)
}
\arguments{
\item{object}{one of \code{jplace}, \code{beast} object}
}
\value{
print info
}
\description{
show method for \code{jplace} instance
}
\examples{
jp <- system.file("extdata", "sample.jplace", package="ggtree")
jp <- read.jplace(jp)
show(jp)
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

