% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_transparent}
\alias{theme_transparent}
\title{theme_transparent}
\usage{
theme_transparent(...)
}
\arguments{
\item{...}{additional parameter to tweak the theme}
}
\value{
ggplot object
}
\description{
transparent background theme
}
\author{
Guangchuang Yu
}

