% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_tree}
\alias{theme_tree}
\title{theme_tree}
\usage{
theme_tree(bgcolor = "white", ...)
}
\arguments{
\item{bgcolor}{set background color, defaults to "white"}

\item{...}{additional parameter}
}
\value{
updated ggplot object with new theme
}
\description{
tree theme
}
\details{
'theme_tree' defines a blank background to display tree
}
\examples{
require(ape)
tr <- rtree(10)
ggtree(tr) + theme_tree()
}
\author{
Guangchuang Yu
}
