% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_tree}
\alias{theme_tree}
\title{theme_tree}
\usage{
theme_tree(bgcolor = "white", fgcolor = "black", ...)
}
\arguments{
\item{bgcolor}{background color}

\item{fgcolor}{foreground color}

\item{...}{additional parameter}
}
\value{
updated ggplot object with new theme
}
\description{
tree theme
}
\examples{
require(ape)
tr <- rtree(10)
ggtree(tr) + theme_tree()
}
\author{
Yu Guangchuang
}
