% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggtree.R
\name{theme_tree2}
\alias{theme_tree2}
\title{theme_tree2}
\usage{
theme_tree2(bgcolor = "white", fgcolor = "black", ...)
}
\arguments{
\item{bgcolor}{background color}

\item{fgcolor}{foreground color}

\item{...}{additional parameter}
}
\value{
updated ggplot object with new theme
}
\description{
tree2 theme
}
\examples{
require(ape)
tr <- rtree(10)
ggtree(tr) + theme_tree2()
}
\author{
Yu Guangchuang
}

