% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewClade.R
\name{viewClade}
\alias{viewClade}
\title{viewClade}
\usage{
viewClade(tree_view, node)
}
\arguments{
\item{tree_view}{full tree view}

\item{node}{internal node number}
}
\value{
clade plot
}
\description{
view a clade of tree
}
\author{
Guangchuang Yu
}

