% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade-functions.R
\name{viewClade}
\alias{viewClade}
\title{viewClade}
\usage{
viewClade(tree_view = NULL, node, xmax_adjust = 0)
}
\arguments{
\item{tree_view}{full tree view}

\item{node}{internal node number}

\item{xmax_adjust}{adjust xmax}
}
\value{
clade plot
}
\description{
view a clade of tree
}
\author{
Guangchuang Yu
}

