% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlim.R
\name{xlim_expand}
\alias{xlim_expand}
\title{xlim_expand}
\usage{
xlim_expand(xlim, panel)
}
\arguments{
\item{xlim}{xlim}

\item{panel}{panel}
}
\value{
updated tree view
}
\description{
expand x axis limits for specific panel
}
\author{
guangchuang yu
}
