% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_plot.R
\name{xlim_tree}
\alias{xlim_tree}
\title{xlim_tree}
\usage{
xlim_tree(xlim)
}
\arguments{
\item{xlim}{xlim, should be of length 2}
}
\value{
updated tree view
}
\description{
set x axis limits for Tree panel
}
\author{
guangchuang yu
}

